Main internal functions for querying the Hub'eau APIs

The function doQueryApi is called by all the function querying the APIs and return the raw data sent by the API. Pagination of the queries is handled automatically and the returned list is the concatenation of all the results sent by the API.

doApiQuery(
  api,
  operation,
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)

get_available_params(
  api,
  operation,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)

Arguments

api
a character name of the API (e.g.: "indicateurs_services", "prelevements"...), see example for available APIs

operation
a character name of the operation, see example for available operations in an API

params
a list the list of parameters of the queries and their values in the format list(ParamName = "Param value", ...), use the function get_available_params for a list of the available parameters for a given operation in an API and see the API documentation for the complete list of available filter parameters

cfg
a config object Configuration of the communication. Use by default the internal package configuration

Value

A list with the concatenated results returned by the API.

Details

The function get_available_params returns the list of available query parameters for a given operation in an API.

The functions get_[api]_[operation] call the function doQueryApi and convert the response in a convenient format for the user (data.frame or tibble::tibble)

Examples

# To get the available APIs in the package
cfg <- config::get(file = system.file("config.yml", package = "hubeau"))
names(cfg$apis)

#> [1] "prelevements"         "indicateurs_services" "hydrometrie"         
#> [4] "niveaux_nappes"      

# To get the available operation in an API
names(cfg$apis[["prelevements"]]$operations)

#> [1] "points_prelevement" "chroniques"         "ouvrages"          

# To get available parameters in operation "chroniques" of the API "prelevements"
get_available_params(api = "prelevements", operation = "chroniques")

#>  [1] "annee"                      "bbox"                      
#>  [3] "code_commune_insee"         "code_departement"          
#>  [5] "code_mode_obtention_volume" "code_ouvrage"              
#>  [7] "code_qualification_volume"  "code_statut_instruction"   
#>  [9] "code_statut_volume"         "code_usage"                
#> [11] "distance"                   "fields"                    
#> [13] "latitude"                   "libelle_departement"       
#> [15] "longitude"                  "nom_commune"               
#> [17] "prelevement_ecrasant"       "producteur_donnee"         
#> [19] "volume_max"                 "volume_min"                

# To query the operation "chroniques" of the API "prelevements"
# on all devices in the commune of Romilly-sur-Seine in 2018
doApiQuery(api = "prelevements",
           operation = "chroniques",
           params = list(code_commune_insee = "10323", annee = "2018"))

#> [[1]]
#> [[1]]$code_ouvrage
#> [1] "OPR0000032603"
#> 
#> [[1]]$annee
#> [1] 2018
#> 
#> [[1]]$volume
#> [1] 402383
#> 
#> [[1]]$code_usage
#> [1] "AEP"
#> 
#> [[1]]$libelle_usage
#> [1] "EAU POTABLE"
#> 
#> [[1]]$code_statut_volume
#> [1] "1"
#> 
#> [[1]]$libelle_statut_volume
#> [1] "Contrôlé Niveau 1"
#> 
#> [[1]]$code_qualification_volume
#> [1] "1"
#> 
#> [[1]]$libelle_qualification_volume
#> [1] "Correcte"
#> 
#> [[1]]$code_statut_instruction
#> [1] "REA"
#> 
#> [[1]]$libelle_statut_instruction
#> [1] "Prélèvement réalisé"
#> 
#> [[1]]$code_mode_obtention_volume
#> [1] "MED"
#> 
#> [[1]]$libelle_mode_obtention_volume
#> [1] "Mesure directe"
#> 
#> [[1]]$prelevement_ecrasant
#> [1] FALSE
#> 
#> [[1]]$producteur_donnee
#> [1] "AESN"
#> 
#> [[1]]$longitude
#> [1] 3.721672
#> 
#> [[1]]$latitude
#> [1] 48.51858
#> 
#> [[1]]$code_commune_insee
#> [1] "10323"
#> 
#> [[1]]$nom_commune
#> [1] "Romilly-sur-Seine"
#> 
#> [[1]]$code_departement
#> [1] "10"
#> 
#> [[1]]$libelle_departement
#> [1] "Aube"
#> 
#> [[1]]$nom_ouvrage
#> [1] "ROMILLY SUR SEINE"
#> 
#> [[1]]$geometry
#> [[1]]$geometry$type
#> [1] "Point"
#> 
#> [[1]]$geometry$crs
#> [[1]]$geometry$crs$type
#> [1] "name"
#> 
#> [[1]]$geometry$crs$properties
#> [[1]]$geometry$crs$properties$name
#> [1] "urn:ogc:def:crs:OGC:1.3:CRS84"
#> 
#> 
#> 
#> [[1]]$geometry$coordinates
#> [[1]]$geometry$coordinates[[1]]
#> [1] 3.721672
#> 
#> [[1]]$geometry$coordinates[[2]]
#> [1] 48.51858
#> 
#> 
#> 
#> [[1]]$uri_ouvrage
#> [1] "https://id.eaufrance.fr/OuvragePrel/OPR0000032603"
#> 
#> 
#> [[2]]
#> [[2]]$code_ouvrage
#> [1] "OPR0000032604"
#> 
#> [[2]]$annee
#> [1] 2018
#> 
#> [[2]]$volume
#> [1] 402383
#> 
#> [[2]]$code_usage
#> [1] "AEP"
#> 
#> [[2]]$libelle_usage
#> [1] "EAU POTABLE"
#> 
#> [[2]]$code_statut_volume
#> [1] "1"
#> 
#> [[2]]$libelle_statut_volume
#> [1] "Contrôlé Niveau 1"
#> 
#> [[2]]$code_qualification_volume
#> [1] "1"
#> 
#> [[2]]$libelle_qualification_volume
#> [1] "Correcte"
#> 
#> [[2]]$code_statut_instruction
#> [1] "REA"
#> 
#> [[2]]$libelle_statut_instruction
#> [1] "Prélèvement réalisé"
#> 
#> [[2]]$code_mode_obtention_volume
#> [1] "MED"
#> 
#> [[2]]$libelle_mode_obtention_volume
#> [1] "Mesure directe"
#> 
#> [[2]]$prelevement_ecrasant
#> [1] FALSE
#> 
#> [[2]]$producteur_donnee
#> [1] "AESN"
#> 
#> [[2]]$longitude
#> [1] 3.721672
#> 
#> [[2]]$latitude
#> [1] 48.51858
#> 
#> [[2]]$code_commune_insee
#> [1] "10323"
#> 
#> [[2]]$nom_commune
#> [1] "Romilly-sur-Seine"
#> 
#> [[2]]$code_departement
#> [1] "10"
#> 
#> [[2]]$libelle_departement
#> [1] "Aube"
#> 
#> [[2]]$nom_ouvrage
#> [1] "ROMILLY SUR SEINE"
#> 
#> [[2]]$geometry
#> [[2]]$geometry$type
#> [1] "Point"
#> 
#> [[2]]$geometry$crs
#> [[2]]$geometry$crs$type
#> [1] "name"
#> 
#> [[2]]$geometry$crs$properties
#> [[2]]$geometry$crs$properties$name
#> [1] "urn:ogc:def:crs:OGC:1.3:CRS84"
#> 
#> 
#> 
#> [[2]]$geometry$coordinates
#> [[2]]$geometry$coordinates[[1]]
#> [1] 3.721672
#> 
#> [[2]]$geometry$coordinates[[2]]
#> [1] 48.51858
#> 
#> 
#> 
#> [[2]]$uri_ouvrage
#> [1] "https://id.eaufrance.fr/OuvragePrel/OPR0000032604"
#> 
#> 
#> [[3]]
#> [[3]]$code_ouvrage
#> [1] "OPR0000040119"
#> 
#> [[3]]$annee
#> [1] 2018
#> 
#> [[3]]$volume
#> [1] 25174
#> 
#> [[3]]$code_usage
#> [1] "IND"
#> 
#> [[3]]$libelle_usage
#> [1] "INDUSTRIE et ACTIVITES ECONOMIQUES (hors irrigation, hors énergie)"
#> 
#> [[3]]$code_statut_volume
#> [1] "1"
#> 
#> [[3]]$libelle_statut_volume
#> [1] "Contrôlé Niveau 1"
#> 
#> [[3]]$code_qualification_volume
#> [1] "1"
#> 
#> [[3]]$libelle_qualification_volume
#> [1] "Correcte"
#> 
#> [[3]]$code_statut_instruction
#> [1] "REA"
#> 
#> [[3]]$libelle_statut_instruction
#> [1] "Prélèvement réalisé"
#> 
#> [[3]]$code_mode_obtention_volume
#> [1] "MED"
#> 
#> [[3]]$libelle_mode_obtention_volume
#> [1] "Mesure directe"
#> 
#> [[3]]$prelevement_ecrasant
#> [1] FALSE
#> 
#> [[3]]$producteur_donnee
#> [1] "AESN"
#> 
#> [[3]]$longitude
#> [1] 3.733233
#> 
#> [[3]]$latitude
#> [1] 48.5101
#> 
#> [[3]]$code_commune_insee
#> [1] "10323"
#> 
#> [[3]]$nom_commune
#> [1] "Romilly-sur-Seine"
#> 
#> [[3]]$code_departement
#> [1] "10"
#> 
#> [[3]]$libelle_departement
#> [1] "Aube"
#> 
#> [[3]]$nom_ouvrage
#> [1] "S.N.C.F"
#> 
#> [[3]]$geometry
#> [[3]]$geometry$type
#> [1] "Point"
#> 
#> [[3]]$geometry$crs
#> [[3]]$geometry$crs$type
#> [1] "name"
#> 
#> [[3]]$geometry$crs$properties
#> [[3]]$geometry$crs$properties$name
#> [1] "urn:ogc:def:crs:OGC:1.3:CRS84"
#> 
#> 
#> 
#> [[3]]$geometry$coordinates
#> [[3]]$geometry$coordinates[[1]]
#> [1] 3.733233
#> 
#> [[3]]$geometry$coordinates[[2]]
#> [1] 48.5101
#> 
#> 
#> 
#> [[3]]$uri_ouvrage
#> [1] "https://id.eaufrance.fr/OuvragePrel/OPR0000040119"
#> 
#> 
#> [[4]]
#> [[4]]$code_ouvrage
#> [1] "OPR0000040120"
#> 
#> [[4]]$annee
#> [1] 2018
#> 
#> [[4]]$volume
#> [1] 27242
#> 
#> [[4]]$code_usage
#> [1] "IND"
#> 
#> [[4]]$libelle_usage
#> [1] "INDUSTRIE et ACTIVITES ECONOMIQUES (hors irrigation, hors énergie)"
#> 
#> [[4]]$code_statut_volume
#> [1] "1"
#> 
#> [[4]]$libelle_statut_volume
#> [1] "Contrôlé Niveau 1"
#> 
#> [[4]]$code_qualification_volume
#> [1] "1"
#> 
#> [[4]]$libelle_qualification_volume
#> [1] "Correcte"
#> 
#> [[4]]$code_statut_instruction
#> [1] "REA"
#> 
#> [[4]]$libelle_statut_instruction
#> [1] "Prélèvement réalisé"
#> 
#> [[4]]$code_mode_obtention_volume
#> [1] "MED"
#> 
#> [[4]]$libelle_mode_obtention_volume
#> [1] "Mesure directe"
#> 
#> [[4]]$prelevement_ecrasant
#> [1] FALSE
#> 
#> [[4]]$producteur_donnee
#> [1] "AESN"
#> 
#> [[4]]$longitude
#> [1] 3.730351
#> 
#> [[4]]$latitude
#> [1] 48.51117
#> 
#> [[4]]$code_commune_insee
#> [1] "10323"
#> 
#> [[4]]$nom_commune
#> [1] "Romilly-sur-Seine"
#> 
#> [[4]]$code_departement
#> [1] "10"
#> 
#> [[4]]$libelle_departement
#> [1] "Aube"
#> 
#> [[4]]$nom_ouvrage
#> [1] "S.N.C.F"
#> 
#> [[4]]$geometry
#> [[4]]$geometry$type
#> [1] "Point"
#> 
#> [[4]]$geometry$crs
#> [[4]]$geometry$crs$type
#> [1] "name"
#> 
#> [[4]]$geometry$crs$properties
#> [[4]]$geometry$crs$properties$name
#> [1] "urn:ogc:def:crs:OGC:1.3:CRS84"
#> 
#> 
#> 
#> [[4]]$geometry$coordinates
#> [[4]]$geometry$coordinates[[1]]
#> [1] 3.730351
#> 
#> [[4]]$geometry$coordinates[[2]]
#> [1] 48.51117
#> 
#> 
#> 
#> [[4]]$uri_ouvrage
#> [1] "https://id.eaufrance.fr/OuvragePrel/OPR0000040120"
#> 
#> 
#> [[5]]
#> [[5]]$code_ouvrage
#> [1] "OPR0000199638"
#> 
#> [[5]]$annee
#> [1] 2018
#> 
#> [[5]]$volume
#> [1] 30588
#> 
#> [[5]]$code_usage
#> [1] "IRR"
#> 
#> [[5]]$libelle_usage
#> [1] "IRRIGATION"
#> 
#> [[5]]$code_statut_volume
#> [1] "1"
#> 
#> [[5]]$libelle_statut_volume
#> [1] "Contrôlé Niveau 1"
#> 
#> [[5]]$code_qualification_volume
#> [1] "1"
#> 
#> [[5]]$libelle_qualification_volume
#> [1] "Correcte"
#> 
#> [[5]]$code_statut_instruction
#> [1] "REA"
#> 
#> [[5]]$libelle_statut_instruction
#> [1] "Prélèvement réalisé"
#> 
#> [[5]]$code_mode_obtention_volume
#> [1] "MED"
#> 
#> [[5]]$libelle_mode_obtention_volume
#> [1] "Mesure directe"
#> 
#> [[5]]$prelevement_ecrasant
#> [1] FALSE
#> 
#> [[5]]$producteur_donnee
#> [1] "AESN"
#> 
#> [[5]]$longitude
#> [1] 3.699283
#> 
#> [[5]]$latitude
#> [1] 48.51774
#> 
#> [[5]]$code_commune_insee
#> [1] "10323"
#> 
#> [[5]]$nom_commune
#> [1] "Romilly-sur-Seine"
#> 
#> [[5]]$code_departement
#> [1] "10"
#> 
#> [[5]]$libelle_departement
#> [1] "Aube"
#> 
#> [[5]]$nom_ouvrage
#> [1] "SCEA FERME DES HAUTS BUISSONS"
#> 
#> [[5]]$geometry
#> [[5]]$geometry$type
#> [1] "Point"
#> 
#> [[5]]$geometry$crs
#> [[5]]$geometry$crs$type
#> [1] "name"
#> 
#> [[5]]$geometry$crs$properties
#> [[5]]$geometry$crs$properties$name
#> [1] "urn:ogc:def:crs:OGC:1.3:CRS84"
#> 
#> 
#> 
#> [[5]]$geometry$coordinates
#> [[5]]$geometry$coordinates[[1]]
#> [1] 3.699283
#> 
#> [[5]]$geometry$coordinates[[2]]
#> [1] 48.51774
#> 
#> 
#> 
#> [[5]]$uri_ouvrage
#> [1] "https://id.eaufrance.fr/OuvragePrel/OPR0000199638"
#> 
#> 
#> [[6]]
#> [[6]]$code_ouvrage
#> [1] "OPR0000333915"
#> 
#> [[6]]$annee
#> [1] 2018
#> 
#> [[6]]$volume
#> [1] 305611
#> 
#> [[6]]$code_usage
#> [1] "IND"
#> 
#> [[6]]$libelle_usage
#> [1] "INDUSTRIE et ACTIVITES ECONOMIQUES (hors irrigation, hors énergie)"
#> 
#> [[6]]$code_statut_volume
#> [1] "1"
#> 
#> [[6]]$libelle_statut_volume
#> [1] "Contrôlé Niveau 1"
#> 
#> [[6]]$code_qualification_volume
#> [1] "1"
#> 
#> [[6]]$libelle_qualification_volume
#> [1] "Correcte"
#> 
#> [[6]]$code_statut_instruction
#> [1] "REA"
#> 
#> [[6]]$libelle_statut_instruction
#> [1] "Prélèvement réalisé"
#> 
#> [[6]]$code_mode_obtention_volume
#> [1] "MED"
#> 
#> [[6]]$libelle_mode_obtention_volume
#> [1] "Mesure directe"
#> 
#> [[6]]$prelevement_ecrasant
#> [1] FALSE
#> 
#> [[6]]$producteur_donnee
#> [1] "AESN"
#> 
#> [[6]]$longitude
#> [1] 3.735902
#> 
#> [[6]]$latitude
#> [1] 48.53133
#> 
#> [[6]]$code_commune_insee
#> [1] "10323"
#> 
#> [[6]]$nom_commune
#> [1] "Romilly-sur-Seine"
#> 
#> [[6]]$code_departement
#> [1] "10"
#> 
#> [[6]]$libelle_departement
#> [1] "Aube"
#> 
#> [[6]]$nom_ouvrage
#> [1] "STE MORGAGNI ZEIMETT"
#> 
#> [[6]]$geometry
#> [[6]]$geometry$type
#> [1] "Point"
#> 
#> [[6]]$geometry$crs
#> [[6]]$geometry$crs$type
#> [1] "name"
#> 
#> [[6]]$geometry$crs$properties
#> [[6]]$geometry$crs$properties$name
#> [1] "urn:ogc:def:crs:OGC:1.3:CRS84"
#> 
#> 
#> 
#> [[6]]$geometry$coordinates
#> [[6]]$geometry$coordinates[[1]]
#> [1] 3.735902
#> 
#> [[6]]$geometry$coordinates[[2]]
#> [1] 48.53133
#> 
#> 
#> 
#> [[6]]$uri_ouvrage
#> [1] "https://id.eaufrance.fr/OuvragePrel/OPR0000333915"
#> 
#>