Retrieve data from API "Ecoulement des cours d'eau"

The data originate from the "ONDE" river low waters monitoring network. Available endpoints are:

get_poisson_stations retrieves sites data and location

get_poisson_observations retrieves flow information

get_poisson_campagnes retrieves annual surveys

See the API documentation for available filter parameters: https://hubeau.eaufrance.fr/page/api-ecoulement

get_ecoulement_stations(
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)

get_ecoulement_observations(
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)

get_ecoulement_campagnes(
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)

Arguments

params

a list the list of parameters of the queries and their values in the format list(ParamName = "Param value", ...), use the function get_available_params for a list of the available query parameters for a given API endpoint and see the API documentation for their descriptions

cfg

a config object describing the configuration of the APIs. Use the internal package configuration by default configuration

Value

a tibble::tibble with all available parameters in columns.

See also

Examples

# Get the query parameters for the requested API/endpoint
get_available_params(api = "ecoulement",
                     endpoint = "observations")
#>  [1] "format"               "code_station"         "libelle_station"     
#>  [4] "code_departement"     "libelle_departement"  "code_commune"        
#>  [7] "libelle_commune"      "code_region"          "libelle_region"      
#> [10] "code_bassin"          "libelle_bassin"       "code_cours_eau"      
#> [13] "libelle_cours_eau"    "code_campagne"        "code_reseau"         
#> [16] "libelle_reseau"       "date_observation_min" "date_observation_max"
#> [19] "code_ecoulement"      "libelle_ecoulement"   "longitude"           
#> [22] "latitude"             "distance"             "bbox"                
#> [25] "page"                 "size"                 "fields"              
#> [28] "accept"              

# Retrieve the river flow data in the Jura departement since 2022-01-01 with
# a selection of the fields
onde_39 <- get_ecoulement_observations(
  list(
     code_departement = "39",
     date_observation_min = "2022-01-01",
     fields = "code_station,libelle_station,date_observation,libelle_ecoulement"
      )
 )