Retrieve data from API "Hydrobiologie"

The data originate from the "NAIADES" database. Available endpoints are:

get_hydrobio_stations_hydrobio retrieves site data and locations

get_hydrobio_indices retrieves bioassessment indices values

get_hydrobio_taxons retrieves taxa data

See the API documentation for available filter parameters: https://hubeau.eaufrance.fr/page/api-hydrobiologie

get_hydrobio_stations_hydrobio(...)

get_hydrobio_indices(...)

get_hydrobio_taxons(...)

Arguments

...

parameters of the queries and their values in the format Param1_Name = "Param1 value", Param2_Name = "Param2 value", use the function list_params for a list of the available filter parameters for a given API endpoint and see the API documentation for their description

Value

A tibble::tibble with one row by record and one column by field.

Examples

# \dontrun{
# Retrieve the hydrobiology monitoring sites in the Pays-de-Loire region
list_params(api = "hydrobio",
            endpoint = "stations_hydrobio")
#>  [1] "bbox"                     "code_bassin"             
#>  [3] "code_commune"             "code_cours_eau"          
#>  [5] "code_departement"         "code_masse_eau"          
#>  [7] "code_region"              "code_sous_bassin"        
#>  [9] "code_station_hydrobio"    "codes_appel_taxons"      
#> [11] "codes_indices"            "codes_reseaux"           
#> [13] "codes_supports"           "date_service"            
#> [15] "distance"                 "fields"                  
#> [17] "format"                   "latitude"                
#> [19] "libelle_station_hydrobio" "longitude"               
#> [21] "page"                     "size"                    
#> [23] "sort"                    

get_hydrobio_stations_hydrobio(code_region = 52)
#> # A tibble: 1,107 × 1,919
#>    code_station_hydrobio libelle_station_hydrobio           uri_station_hydrobio
#>    <chr>                 <chr>                              <chr>               
#>  1 03271654              LA CHEVAUCHERIE A MONTAUDIN 1      https://id.eaufranc…
#>  2 03271657              LE RUISSEAU DE LA MORINIÈRE A SAI… https://id.eaufranc…
#>  3 03271661              LA HOGUE A SAINT-BERTHEVIN-LA-TAN… https://id.eaufranc…
#>  4 03271679              LA FUTAIE A SAINT-ELLIER-DU-MAINE… https://id.eaufranc…
#>  5 03271699              LA FUTAIE A PONTMAIN 2             https://id.eaufranc…
#>  6 03271710              LA FUTAIE A PONTMAIN 3             https://id.eaufranc…
#>  7 03271730              LA GLAINE A LA BAZOUGE-DU-DESERT 2 https://id.eaufranc…
#>  8 03271850              L'AIRON A LANDIVY 1                https://id.eaufranc…
#>  9 04098400              LOIRE à MONTSOREAU                 https://id.eaufranc…
#> 10 04102000              THOUET à MONTREUIL-BELLAY          https://id.eaufranc…
#> # ℹ 1,097 more rows
#> # ℹ 1,916 more variables: coordonnee_x <dbl>, coordonnee_y <dbl>,
#> #   code_projection <chr>, code_cours_eau <chr>, libelle_cours_eau <chr>,
#> #   uri_cours_eau <chr>, code_masse_eau <chr>, libelle_masse_eau <chr>,
#> #   uri_masse_eau <chr>, code_sous_bassin <chr>, libelle_sous_bassin <chr>,
#> #   code_bassin <chr>, libelle_bassin <chr>, code_commune <chr>,
#> #   libelle_commune <chr>, code_departement <chr>, libelle_departement <chr>, …

# Retrieve the hydrobiological bioassessment indices in the city of Rennes
list_params(api = "hydrobio",
            endpoint = "indices")
#>  [1] "bbox"                       "code_banque_reference"     
#>  [3] "code_bassin"                "code_commune"              
#>  [5] "code_cours_eau"             "code_departement"          
#>  [7] "code_indice"                "code_masse_eau"            
#>  [9] "code_operation_prelevement" "code_prelevement"          
#> [11] "code_qualification"         "code_region"               
#> [13] "code_sous_bassin"           "code_station_hydrobio"     
#> [15] "code_support"               "date_debut_prelevement"    
#> [17] "date_fin_prelevement"       "distance"                  
#> [19] "fields"                     "format"                    
#> [21] "latitude"                   "libelle_indice"            
#> [23] "longitude"                  "page"                      
#> [25] "resultat_indice_max"        "resultat_indice_min"       
#> [27] "size"                       "sort"                      

get_hydrobio_indices(code_commune = 35051)
#> # A tibble: 655 × 44
#>    code_indice libelle_indice       code_station_hydrobio libelle_station_hydr…¹
#>    <chr>       <chr>                <chr>                 <chr>                 
#>  1 9126        Note en EQR de l'In… 04204300              VILAINE à CESSON-SEVI…
#>  2 9143        Métrique de composi… 04204300              VILAINE à CESSON-SEVI…
#>  3 9142        Métrique de composi… 04204300              VILAINE à CESSON-SEVI…
#>  4 9141        Métrique de composi… 04204300              VILAINE à CESSON-SEVI…
#>  5 9140        Métrique de Biomass… 04204300              VILAINE à CESSON-SEVI…
#>  6 9140        Métrique de Biomass… 04204300              VILAINE à CESSON-SEVI…
#>  7 9126        Note en EQR de l'In… 04204300              VILAINE à CESSON-SEVI…
#>  8 9143        Métrique de composi… 04204300              VILAINE à CESSON-SEVI…
#>  9 9142        Métrique de composi… 04204300              VILAINE à CESSON-SEVI…
#> 10 9141        Métrique de composi… 04204300              VILAINE à CESSON-SEVI…
#> # ℹ 645 more rows
#> # ℹ abbreviated name: ¹​libelle_station_hydrobio
#> # ℹ 40 more variables: uri_station_hydrobio <chr>, date_prelevement <chr>,
#> #   resultat_indice <dbl>, unite_indice <chr>, coordonnee_x <dbl>,
#> #   coordonnee_y <dbl>, code_projection <chr>, code_cours_eau <chr>,
#> #   libelle_cours_eau <chr>, uri_cours_eau <chr>, code_masse_eau <chr>,
#> #   libelle_masse_eau <lgl>, uri_masse_eau <chr>, code_sous_bassin <lgl>, …

# species records in the city of Rennes
list_params(api = "hydrobio",
            endpoint = "taxons")
#>  [1] "bbox"                       "code_appel_taxon"          
#>  [3] "code_banque_reference"      "code_bassin"               
#>  [5] "code_commune"               "code_cours_eau"            
#>  [7] "code_departement"           "code_masse_eau"            
#>  [9] "code_operation_prelevement" "code_prelevement"          
#> [11] "code_qualification"         "code_region"               
#> [13] "code_sous_bassin"           "code_station_hydrobio"     
#> [15] "code_support"               "code_type_resultat"        
#> [17] "codes_indices_operation"    "date_debut_prelevement"    
#> [19] "date_fin_prelevement"       "distance"                  
#> [21] "fields"                     "format"                    
#> [23] "latitude"                   "longitude"                 
#> [25] "page"                       "size"                      
#> [27] "sort"                      

get_hydrobio_taxons(code_commune = 35051)
#> # A tibble: 14,196 × 90
#>    code_station_hydrobio libelle_station_hydrobio uri_station_hydrobio          
#>    <chr>                 <chr>                    <chr>                         
#>  1 04204300              VILAINE à CESSON-SEVIGNE https://id.eaufrance.fr/Stati…
#>  2 04204300              VILAINE à CESSON-SEVIGNE https://id.eaufrance.fr/Stati…
#>  3 04204300              VILAINE à CESSON-SEVIGNE https://id.eaufrance.fr/Stati…
#>  4 04204300              VILAINE à CESSON-SEVIGNE https://id.eaufrance.fr/Stati…
#>  5 04204300              VILAINE à CESSON-SEVIGNE https://id.eaufrance.fr/Stati…
#>  6 04204300              VILAINE à CESSON-SEVIGNE https://id.eaufrance.fr/Stati…
#>  7 04204300              VILAINE à CESSON-SEVIGNE https://id.eaufrance.fr/Stati…
#>  8 04204300              VILAINE à CESSON-SEVIGNE https://id.eaufrance.fr/Stati…
#>  9 04204300              VILAINE à CESSON-SEVIGNE https://id.eaufrance.fr/Stati…
#> 10 04204300              VILAINE à CESSON-SEVIGNE https://id.eaufrance.fr/Stati…
#> # ℹ 14,186 more rows
#> # ℹ 87 more variables: date_prelevement <chr>, code_support <chr>,
#> #   libelle_support <chr>, code_appel_taxon <chr>, libelle_appel_taxon <chr>,
#> #   codes_taxons_parents1 <chr>, codes_taxons_parents2 <chr>,
#> #   codes_taxons_parents3 <chr>, codes_taxons_parents4 <chr>,
#> #   codes_taxons_parents5 <chr>, codes_taxons_parents6 <chr>,
#> #   codes_taxons_parents7 <chr>, codes_taxons_parents8 <chr>, …
# }