Retrieve data from API "Hydrométrie"

Available endpoints are:

get_hydrometrie_obs_elab retrieves hydrometric elaborate observations (daily/monthly mean flow)

get_hydrometrie_observations_tr retrieves hydrometric "real time" observations ()

get_hydrometrie_sites retrieves hydrometric sites

get_hydrometrie_stations retrieves hydrometric stations

See the API documentation of each endpoint for available filter parameters: https://hubeau.eaufrance.fr/page/api-hydrometrie

get_hydrometrie_obs_elab(
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)

get_hydrometrie_observations_tr(
  params,
  entities = "station",
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)

get_hydrometrie_sites(
  params,
  unique_site = TRUE,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)

get_hydrometrie_stations(
  params,
  code_sandre_reseau_station = FALSE,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)

Arguments

params

a list the list of parameters of the queries and their values in the format list(ParamName = "Param value", ...), use the function get_available_params for a list of the available query parameters for a given API endpoint and see the API documentation for their description

cfg

a config object describing the configuration of the APIs. Use the internal package configuration by default configuration

entities

1-length character string filtering the rows of the returned value, possible values are: "station" for filtering on station rows, "site" for filtering on site rows, "both" for keeping all the rows

unique_site

optional logical, if set to FALSE sites with several different locations produce one row by different location otherwise the first location found is used for fields code_commune_site, libelle_commune, code_departement, code_region, libelle_region, libelle_departement

code_sandre_reseau_station

optional logical indicating if code_sandre_reseau_station field is included in the result; if so, one line is added by item and other fields are repeated

Value

a tibble::tibble with all available parameters in columns.

See also

Examples

# Retrieve the hydrometric sites in the department of Aube
get_hydrometrie_sites(list(code_departement = "10"))
#> Warning: The site 'H0203020' has 2 different locations, only the first one is returned
#> # A tibble: 27 × 34
#>    code_site libelle_site            type_site coordonnee_x_si… coordonnee_y_si…
#>    <chr>     <chr>                   <chr>                <dbl>            <dbl>
#>  1 F3550001  La Vanne à Paisy-Cosdon REEL                752416          6793820
#>  2 H0100020  La Seine à Plaines-Sai… REEL                810441          6767194
#>  3 H0203020  La Laignes aux Riceys … REEL                801571          6763252
#>  4 H0210010  La Seine à Polisy       REEL                802885          6774610
#>  5 H0321040  L'Ource à Celles-sur-O… REEL                804893          6776215
#>  6 H0400010  La Seine à Bar-sur-Sei… REEL                802511          6780502
#>  7 H0400020  La Seine à Courtenot    REEL                799402          6782747
#>  8 H0413010  La Sarce à Virey-sous-… REEL                796857          6783149
#>  9 H0503010  L'Hozain à Buchères [C… REEL                782364          6794544
#> 10 H0702010  La Barse à Montiéramey  REEL                795978          6793759
#> # … with 17 more rows, and 29 more variables: code_projection <int>,
#> #   longitude_site <dbl>, latitude_site <dbl>, altitude_site <dbl>,
#> #   code_systeme_alti_site <int>, surface_bv <dbl>, statut_site <int>,
#> #   premier_mois_etiage_site <int>, premier_mois_annee_hydro_site <int>,
#> #   influence_generale_site <int>, code_entite_hydro_site <chr>,
#> #   code_troncon_hydro_site <chr>, code_commune_site <chr>,
#> #   code_zone_hydro_site <chr>, libelle_commune <chr>, …

# The same operation returning 2 rows for the site 'H0203020' which has 2 different locations
get_hydrometrie_sites(list(code_departement = "10"), unique_site = FALSE)
#> # A tibble: 28 × 34
#>    code_site libelle_site            type_site coordonnee_x_si… coordonnee_y_si…
#>    <chr>     <chr>                   <chr>                <dbl>            <dbl>
#>  1 F3550001  La Vanne à Paisy-Cosdon REEL                752416          6793820
#>  2 H0100020  La Seine à Plaines-Sai… REEL                810441          6767194
#>  3 H0203020  La Laignes aux Riceys … REEL                801571          6763252
#>  4 H0203020  La Laignes aux Riceys … REEL                801571          6763252
#>  5 H0210010  La Seine à Polisy       REEL                802885          6774610
#>  6 H0321040  L'Ource à Celles-sur-O… REEL                804893          6776215
#>  7 H0400010  La Seine à Bar-sur-Sei… REEL                802511          6780502
#>  8 H0400020  La Seine à Courtenot    REEL                799402          6782747
#>  9 H0413010  La Sarce à Virey-sous-… REEL                796857          6783149
#> 10 H0503010  L'Hozain à Buchères [C… REEL                782364          6794544
#> # … with 18 more rows, and 29 more variables: code_projection <int>,
#> #   longitude_site <dbl>, latitude_site <dbl>, altitude_site <dbl>,
#> #   code_systeme_alti_site <int>, surface_bv <dbl>, statut_site <int>,
#> #   premier_mois_etiage_site <int>, premier_mois_annee_hydro_site <int>,
#> #   influence_generale_site <int>, code_entite_hydro_site <chr>,
#> #   code_troncon_hydro_site <chr>, code_commune_site <chr>,
#> #   code_zone_hydro_site <chr>, libelle_commune <chr>, …

if (FALSE) {
# This function is currently (2021-12-23) unstable and can be unavailable
# (See https://github.com/BRGM/hubeau/issues/85)
# Retrieve the hydrometric stations in the department of Aube
get_hydrometrie_stations(list(code_departement = "10"))
}

# Which parameters are available for endpoint "obs_elab" of API "hydrometrie"?
get_available_params("hydrometrie", "obs_elab")
#>  [1] "bbox"                "code_entite"         "cursor"             
#>  [4] "date_debut_obs_elab" "date_fin_obs_elab"   "distance"           
#>  [7] "fields"              "grandeur_hydro_elab" "latitude"           
#> [10] "longitude"           "resultat_max"        "resultat_min"       

# Retrieve the hydrometric monthly mean flow at site 'H0203020'
get_hydrometrie_obs_elab(list(code_entite = "H0203020", grandeur_hydro_elab = "QmM"))
#> # A tibble: 651 × 14
#>    code_site code_station date_obs_elab resultat_obs_elab date_prod  code_statut
#>    <chr>     <chr>        <chr>                     <dbl> <chr>            <int>
#>  1 H0203020  H020302001   1968-02-01                 6246 2021-09-1…          16
#>  2 H0203020  H020302001   1968-03-01                 3578 2021-09-1…          16
#>  3 H0203020  H020302001   1968-04-01                 1993 2021-09-1…          16
#>  4 H0203020  H020302001   1968-05-01                 4934 2021-09-1…          16
#>  5 H0203020  H020302001   1968-06-01                 1610 2021-09-1…          16
#>  6 H0203020  H020302001   1968-07-01                 1022 2021-09-1…          16
#>  7 H0203020  H020302001   1968-08-01                 1040 2021-09-1…          16
#>  8 H0203020  H020302001   1968-09-01                 3451 2021-09-1…          16
#>  9 H0203020  H020302001   1968-10-01                 4242 2021-09-1…          16
#> 10 H0203020  H020302001   1968-11-01                 2675 2021-09-1…          16
#> # … with 641 more rows, and 8 more variables: libelle_statut <chr>,
#> #   code_methode <int>, libelle_methode <chr>, code_qualification <int>,
#> #   libelle_qualification <chr>, longitude <dbl>, latitude <dbl>,
#> #   grandeur_hydro_elab <chr>

# Retrieve the hydrometric daily mean flow at site 'H0203020' of the last 30 days
get_hydrometrie_obs_elab(
  list(code_entite = "H0203020",
  date_debut_obs_elab = format(Sys.Date() -30, "%Y-%m-%d"),
  grandeur_hydro_elab = "QmJ"))
#> # A tibble: 22 × 14
#>    code_site code_station date_obs_elab resultat_obs_elab date_prod  code_statut
#>    <chr>     <chr>        <chr>                     <dbl> <chr>            <int>
#>  1 H0203020  H020302002   2022-06-05                 1305 2022-06-0…           4
#>  2 H0203020  H020302002   2022-06-06                 1319 2022-06-0…           4
#>  3 H0203020  H020302002   2022-06-07                 1328 2022-06-0…           4
#>  4 H0203020  H020302002   2022-06-08                 1305 2022-06-0…           4
#>  5 H0203020  H020302002   2022-06-09                 1390 2022-06-1…           4
#>  6 H0203020  H020302002   2022-06-10                 1309 2022-06-1…           4
#>  7 H0203020  H020302002   2022-06-11                 1270 2022-06-1…           4
#>  8 H0203020  H020302002   2022-06-12                 1242 2022-06-1…           4
#>  9 H0203020  H020302002   2022-06-13                 1221 2022-06-1…           4
#> 10 H0203020  H020302002   2022-06-21                 1103 2022-06-2…           4
#> # … with 12 more rows, and 8 more variables: libelle_statut <chr>,
#> #   code_methode <int>, libelle_methode <chr>, code_qualification <int>,
#> #   libelle_qualification <chr>, longitude <dbl>, latitude <dbl>,
#> #   grandeur_hydro_elab <chr>