Retrieve hydrometric "real time" observations from API "Hydrométrie"

See the API documentation for available filter parameters: https://hubeau.eaufrance.fr/page/api-hydrometrie

get_hydrometrie_observations_tr(
  params,
  entities = "station",
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)

Arguments

params

a list the list of parameters of the queries and their values in the format list(ParamName = "Param value", ...), use the function get_available_params for a list of the available parameters for a given operation in an API and see the API documentation for the complete list of available filter parameters

entities

1-character string filtering the rows of the returned value, possible values are: "station" for filtering on station rows, "site" for filtering on site rows, "both" for keeping all the rows

cfg

a config object Configuration of the communication. Use by default the internal package configuration

Value

a tibble::tibble with all available parameters in columns and one row by time step and by station. If the query returns no records then the returned value is NULL.

Examples

# Retrieve the hydrometric last real time observed discharge at station 'H020302002'
get_hydrometrie_observations_tr(list(code_entite = "H0203020", grandeur_hydro = "Q"))
#> # A tibble: 4,287 × 16
#>    code_site code_station grandeur_hydro date_debut_serie     date_fin_serie    
#>    <chr>     <chr>        <chr>          <chr>                <chr>             
#>  1 H0203020  H020302002   Q              2021-12-23T00:00:01Z 2021-12-23T16:00:…
#>  2 H0203020  H020302002   Q              2021-12-23T00:00:01Z 2021-12-23T16:00:…
#>  3 H0203020  H020302002   Q              2021-12-23T00:00:01Z 2021-12-23T16:00:…
#>  4 H0203020  H020302002   Q              2021-12-23T00:00:01Z 2021-12-23T16:00:…
#>  5 H0203020  H020302002   Q              2021-12-23T00:00:01Z 2021-12-23T16:00:…
#>  6 H0203020  H020302002   Q              2021-12-23T00:00:01Z 2021-12-23T16:00:…
#>  7 H0203020  H020302002   Q              2021-12-23T00:00:01Z 2021-12-23T15:00:…
#>  8 H0203020  H020302002   Q              2021-12-23T00:00:01Z 2021-12-23T15:00:…
#>  9 H0203020  H020302002   Q              2021-12-23T00:00:01Z 2021-12-23T15:00:…
#> 10 H0203020  H020302002   Q              2021-12-23T00:00:01Z 2021-12-23T15:00:…
#> # … with 4,277 more rows, and 11 more variables: statut_serie <int>,
#> #   code_systeme_alti_serie <int>, date_obs <chr>, resultat_obs <dbl>,
#> #   code_methode_obs <int>, libelle_methode_obs <chr>,
#> #   code_qualification_obs <int>, libelle_qualification_obs <chr>,
#> #   continuite_obs_hydro <lgl>, longitude <dbl>, latitude <dbl>