Retrieve performance indicators from drinking water supply and sanitation services at commune level

See the API documentation for available filter parameters: https://hubeau.eaufrance.fr/page/api-indicateurs-services

get_indicateurs_services_communes(
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)

Arguments

params

a list the list of parameters of the queries and their values in the format list(ParamName = "Param value", ...), use the function get_available_params for a list of the available parameters for a given operation in an API and see the API documentation for the complete list of available filter parameters

cfg

a config object Configuration of the communication. Use by default the internal package configuration

Value

a tibble::tibble with one row by commune, by service and by year and the following columns:

"code_commune_insee": character identifier of the commune

"nom_commune": character name of the commune

"codes_service": integer identifier of the drinking water supply and/or sanitation service

"annee": integer year of the data

The following columns are the performance indicators flagged by their respective codes. The documentation of these codes can be found at this URL: https://www.services.eaufrance.fr/indicateurs/indicateurs.

Examples

# Retrieve performance indicator time series of the services in Romilly-sur-Seine
get_indicateurs_services_communes(list(code_commune = "10323"))
#> # A tibble: 28 × 39
#>    code_commune_insee nom_commune codes_service noms_service annee D102.0 D151.0
#>    <chr>              <chr>               <int> <chr>        <int>  <dbl>  <dbl>
#>  1 10323              Romilly-su…        148782 eau potable   2018   1.91      1
#>  2 10323              Romilly-su…        148782 eau potable   2017   2.07      1
#>  3 10323              Romilly-su…        148783 assainissem…  2017  NA        NA
#>  4 10323              NA                 148783 assainissem…  2017  NA        NA
#>  5 10323              Romilly-su…        148782 eau potable   2016   2.18      1
#>  6 10323              Romilly-su…        148783 assainissem…  2016  NA        NA
#>  7 10323              NA                 148783 assainissem…  2016  NA        NA
#>  8 10323              Romilly-su…        148782 eau potable   2015   2.2       1
#>  9 10323              NA                 148783 assainissem…  2015  NA        NA
#> 10 10323              Romilly-su…        148783 assainissem…  2015  NA        NA
#> # … with 18 more rows, and 32 more variables: P101.1 <dbl>, P101.1a <lgl>,
#> #   P101.1b <lgl>, P102.1 <dbl>, P102.1a <lgl>, P102.1b <lgl>, P104.3 <dbl>,
#> #   P106.3 <dbl>, P107.2 <dbl>, P108.3 <dbl>, P151.1 <dbl>, P152.1 <dbl>,
#> #   P155.1 <dbl>, VP.177 <dbl>, VP.190 <dbl>, VP.178 <dbl>, VP.191 <dbl>,
#> #   VP.179 <dbl>, VP.061 <dbl>, DC.192 <dbl>, VP.228 <dbl>, VP.231 <dbl>,
#> #   D204.0 <dbl>, P201.1 <dbl>, P203.3 <dbl>, P204.3 <dbl>, P205.3 <dbl>,
#> #   P258.1 <dbl>, VP.186 <dbl>, VP.176 <dbl>, VP.023 <dbl>, VP.229 <dbl>