Retrieve performance indicators from drinking water supply and sanitation services at commune level

See the API documentation for available filter parameters: https://hubeau.eaufrance.fr/page/api-indicateurs-services

get_indicateurs_services_indicateurs(
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)

Arguments

params
a list the list of parameters of the queries and their values in the format list(ParamName = "Param value", ...), use the function get_available_params for a list of the available parameters for a given operation in an API and see the API documentation for the complete list of available filter parameters

cfg
a config object Configuration of the communication. Use by default the internal package configuration

Value

a tibble::tibble with one row by commune, by service and by year and the following columns:

"code_commune_insee": character identifier of the commune

"nom_commune": character name of the commune

"codes_service": integer identifier of the drinking water supply and/or sanitation service

"annee": integer year of the data

The following columns are the performance indicators flagged by their respective codes. The documentation of these codes can be found at this URL: https://www.services.eaufrance.fr/indicateurs/indicateurs.

Examples

# Retrieve the withdrawal time series of the devices located in Romilly-sur-Seine
get_indicateurs_services_indicateurs(list(code_indicateur = "D102.0", annee = "2012"))

#> Error: `map_df()` requires dplyr