Retrieve data from API "Poisson"

Available endpoint:

get_poisson_observations retrieves data of scientific fishery operations

See the API documentation for available filter parameters: https://hubeau.eaufrance.fr/page/api-poisson

get_poisson_observations(
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)

Arguments

params

a list the list of parameters of the queries and their values in the format list(ParamName = "Param value", ...), use the function get_available_params for a list of the available query parameters for a given API endpoint and see the API documentation for their descriptions

cfg

a config object describing the configuration of the APIs. Use the internal package configuration by default configuration

Value

a tibble::tibble with all available parameters in columns.

See also

Examples

# Get the query parameters for the requested API/endpoint
get_available_params(api = "poisson",
                     endpoint = "observations")
#>   [1] "code_operation"                               
#>   [2] "date_operation"                               
#>   [3] "etat_avancement_operation"                    
#>   [4] "code_qualification_operation"                 
#>   [5] "libelle_qualification_operation"              
#>   [6] "code_station"                                 
#>   [7] "libelle_station"                              
#>   [8] "uri_station"                                  
#>   [9] "coordonnee_x_station"                         
#>  [10] "coordonnee_y_station"                         
#>  [11] "code_epsg_projection_station"                 
#>  [12] "code_projection_station"                      
#>  [13] "libelle_projection_station"                   
#>  [14] "code_point_prelevement"                       
#>  [15] "code_point_prelevement_aspe"                  
#>  [16] "libelle_point_prelevement_aspe"               
#>  [17] "coordonnee_x_point_prelevement"               
#>  [18] "coordonnee_y_point_prelevement"               
#>  [19] "code_epsg_projection_point_prelevement"       
#>  [20] "code_projection_point_prelevement"            
#>  [21] "libelle_projection_point_prelevement"         
#>  [22] "lieu_dit_point_prelevement"                   
#>  [23] "localisation_precise_point_prelevement"       
#>  [24] "longitude"                                    
#>  [25] "code_entite_hydro"                            
#>  [26] "latitude"                                     
#>  [27] "libelle_entite_hydro"                         
#>  [28] "uri_entite_hydro"                             
#>  [29] "code_commune"                                 
#>  [30] "libelle_commune"                              
#>  [31] "code_departement"                             
#>  [32] "libelle_departement"                          
#>  [33] "code_region"                                  
#>  [34] "libelle_region"                               
#>  [35] "code_bassin"                                  
#>  [36] "libelle_bassin"                               
#>  [37] "codes_dispositifs_collecte"                   
#>  [38] "libelles_dispositifs_collecte"                
#>  [39] "uris_dispositifs_collecte"                    
#>  [40] "protocole_peche"                              
#>  [41] "objectifs_operation"                          
#>  [42] "code_prelevement_elementaire"                 
#>  [43] "type_prelevement_elementaire"                 
#>  [44] "duree_prelevement_elementaire"                
#>  [45] "numero_passage"                               
#>  [46] "type_points"                                  
#>  [47] "nombre_points"                                
#>  [48] "nombre_points_sans_poisson"                   
#>  [49] "nombre_points_facies_courant"                 
#>  [50] "nombre_points_facies_plat"                    
#>  [51] "nombre_points_facies_profond"                 
#>  [52] "nombre_points_annexes"                        
#>  [53] "nombre_points_berge"                          
#>  [54] "nombre_points_chenal"                         
#>  [55] "nom_ambiance"                                 
#>  [56] "longueur_ambiance"                            
#>  [57] "largeur_ambiance"                             
#>  [58] "profondeur_ambiance"                          
#>  [59] "code_facies_ambiance"                         
#>  [60] "libelle_facies_ambiance"                      
#>  [61] "situation_ambiance"                           
#>  [62] "localisation_ambiance"                        
#>  [63] "classe_vitesse_courant_ambiance"              
#>  [64] "code_ombrage_ambiance"                        
#>  [65] "libelle_ombrage_ambiance"                     
#>  [66] "code_granulo_ambiance_dominant"               
#>  [67] "libelle_granulo_ambiance_dominant"            
#>  [68] "code_granulo_ambiance_accessoire"             
#>  [69] "libelle_granulo_ambiance_accessoire"          
#>  [70] "vegetation_ambiance"                          
#>  [71] "abondance_cache_vegetation_aquatique_ambiance"
#>  [72] "abondance_cache_embacles_ambiance"            
#>  [73] "abondance_cache_vegetation_bordure_ambiance"  
#>  [74] "abondance_cache_sous_berge_ambiance"          
#>  [75] "abondance_cache_granulo_ambiance"             
#>  [76] "observation_ambiance"                         
#>  [77] "code_lot"                                     
#>  [78] "code_type_lot"                                
#>  [79] "libelle_type_lot"                             
#>  [80] "effectif_lot"                                 
#>  [81] "code_taxon"                                   
#>  [82] "code_alternatif_taxon"                        
#>  [83] "nom_commun_taxon"                             
#>  [84] "nom_latin_taxon"                              
#>  [85] "uri_taxon"                                    
#>  [86] "taille_min_lot"                               
#>  [87] "taille_max_lot"                               
#>  [88] "type_longueur_lot"                            
#>  [89] "poids_lot_mesure"                             
#>  [90] "poids_lot_estime"                             
#>  [91] "methode_estimation_poids_lot"                 
#>  [92] "code_individu"                                
#>  [93] "taille_individu"                              
#>  [94] "type_longueur_individu"                       
#>  [95] "mesure_individu_reelle"                       
#>  [96] "poids_individu_mesure"                        
#>  [97] "poids_individu_estime"                        
#>  [98] "methode_estimation_poids_individu"            
#>  [99] "sexe_individu"                                
#> [100] "age_individu"                                 
#> [101] "codes_pathologies_lot"                        
#> [102] "libelles_pathologies_lot"                     
#> [103] "codes_fractions_pathologies_lot"              
#> [104] "libelles_fractions_pathologies_lot"           
#> [105] "uris_fractions_pathologies_lot"               
#> [106] "importances_pathologies_lot"                  
#> [107] "codes_pathologies_individu"                   
#> [108] "libelles_pathologies_individu"                
#> [109] "codes_fractions_pathologies_individu"         
#> [110] "libelles_fractions_pathologies_individu"      
#> [111] "uris_fractions_pathologies_individu"          
#> [112] "importances_pathologies_individu"             
#> [113] "fields"                                       

# Retrieve selected fields on a river fish sampled in Brest
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
fields <- paste("code_operation",
                "date_operation",
                "libelle_point_prelevement_aspe",
                "effectif_lot",
                "code_alternatif_taxon",
                sep = ",")
brest_fishes <- get_poisson_observations(
  list(
    libelle_commune = "Brest",
    fields = fields
    )
 ) %>%
 group_by_at(vars(-effectif_lot)) %>%
   summarise(nb_individals = sum(effectif_lot))
#> `summarise()` has grouped output by 'code_operation', 'date_operation',
#> 'libelle_point_prelevement_aspe'. You can override using the `.groups`
#> argument.

brest_fishes
#> # A tibble: 5 × 5
#> # Groups:   code_operation, date_operation, libelle_point_prelevement_aspe [1]
#>   code_operation date_operation       libelle_point_prelevemen…¹ code_…² nb_in…³
#>   <chr>          <chr>                <chr>                      <chr>     <int>
#> 1 6003           1995-10-12T14:30:00Z La Penfeld à Brest         ANG           3
#> 2 6003           1995-10-12T14:30:00Z La Penfeld à Brest         CHA           1
#> 3 6003           1995-10-12T14:30:00Z La Penfeld à Brest         EPI           1
#> 4 6003           1995-10-12T14:30:00Z La Penfeld à Brest         LOF           2
#> 5 6003           1995-10-12T14:30:00Z La Penfeld à Brest         TRF          17
#> # … with abbreviated variable names ¹​libelle_point_prelevement_aspe,
#> #   ²​code_alternatif_taxon, ³​nb_individals