Retrieve data from API "Prélèvements en eau"

Available endpoints are:

get_prelevements_points_prelevement retrieves withdrawal points

get_prelevements_ouvrages retrieves withdrawal devices

get_prelevements_chroniques retrieves time series of withdrawals

See the API documentation for available filter parameters: https://hubeau.eaufrance.fr/page/api-prelevements-eau

get_prelevements_points_prelevement(...)

get_prelevements_ouvrages(...)

get_prelevements_chroniques(...)

Arguments

...

parameters of the queries and their values in the format Param1_Name = "Param1 value", Param2_Name = "Param2 value", use the function list_params for a list of the available filter parameters for a given API endpoint and see the API documentation for their description

Value

A tibble::tibble with one row by record and one column by field.

Examples

# \dontrun{
# Retrieve the withdrawal points located in Romilly-sur-Seine
get_prelevements_points_prelevement(code_commune_insee = "10323")
#> # A tibble: 9 × 28
#>   code_point_p…¹ nom_p…² date_…³ date_…⁴ code_…⁵ libel…⁶ code_…⁷ libel…⁸ lieu_…⁹
#>   <chr>          <lgl>   <chr>   <lgl>   <chr>   <chr>   <chr>   <chr>   <lgl>  
#> 1 PTP0000000000… NA      1987-0… NA      SOUT    Souter… P       PHYSIQ… NA     
#> 2 PTP0000000000… NA      1987-0… NA      SOUT    Souter… P       PHYSIQ… NA     
#> 3 PTP0000000000… NA      1900-0… NA      SOUT    Souter… P       PHYSIQ… NA     
#> 4 PTP0000000000… NA      1900-0… NA      SOUT    Souter… P       PHYSIQ… NA     
#> 5 PTP0000000000… NA      2006-0… NA      SOUT    Souter… P       PHYSIQ… NA     
#> 6 PTP0000000002… NA      1987-0… NA      SOUT    Souter… P       PHYSIQ… NA     
#> 7 PTP0000000004… NA      2016-0… NA      CONT    Surfac… P       PHYSIQ… NA     
#> 8 PTP0000000007… NA      1900-0… NA      SOUT    Souter… P       PHYSIQ… NA     
#> 9 PTP0000000007… NA      1900-0… NA      SOUT    Souter… P       PHYSIQ… NA     
#> # … with 19 more variables: commentaire <lgl>, code_commune_insee <chr>,
#> #   nom_commune <chr>, code_departement <chr>, libelle_departement <chr>,
#> #   code_entite_hydro_cours_eau <lgl>, uri_entite_hydro_cours_eau <lgl>,
#> #   code_entite_hydro_plan_eau <lgl>, uri_entite_hydro_plan_eau <lgl>,
#> #   code_zone_hydro <chr>, uri_zone_hydro <chr>, code_mer_ocean <lgl>,
#> #   nappe_accompagnement <lgl>, uri_bss_point_eau <chr>, code_ouvrage <chr>,
#> #   uri_ouvrage <chr>, code_bdlisa <lgl>, uri_bdlisa <lgl>, …

# Retrieve the withdrawal devices located in Romilly-sur-Seine
get_prelevements_ouvrages(code_commune_insee = "10323")
#> # A tibble: 9 × 33
#>   code_ouvrage  nom_ou…¹ id_lo…² date_…³ date_…⁴ code_…⁵ libel…⁶ comme…⁷ code_…⁸
#>   <chr>         <chr>    <chr>   <chr>   <lgl>   <chr>   <chr>   <lgl>   <chr>  
#> 1 OPR0000032603 ROMILLY… 02612X… 1987-0… NA      5       Coordo… NA      10323  
#> 2 OPR0000032604 ROMILLY… 02612X… 1987-0… NA      5       Coordo… NA      10323  
#> 3 OPR0000038567 SCEA FE… BSS954… 1900-0… NA      0       Précis… NA      10323  
#> 4 OPR0000040119 S.N.C.F  BSS005… 1900-0… NA      0       Précis… NA      10323  
#> 5 OPR0000040120 S.N.C.F  BSS005… 2006-0… NA      0       Précis… NA      10323  
#> 6 OPR0000199638 SCEA FE… BSS954… 1987-0… NA      0       Précis… NA      10323  
#> 7 OPR0000333915 Soc des… 32055_… 2016-0… NA      0       Précis… NA      10323  
#> 8 OPR0000586224 SCEA DU… BSS952… 1900-0… NA      0       Précis… NA      10323  
#> 9 OPR0000608675 CRBPE R… 31132_… 1900-0… NA      0       Précis… NA      10323  
#> # … with 24 more variables: nom_commune <chr>, code_departement <chr>,
#> #   libelle_departement <chr>, code_type_milieu <chr>,
#> #   libelle_type_milieu <chr>, code_entite_hydro_cours_eau <chr>,
#> #   uri_entite_hydro_cours_eau <chr>, code_entite_hydro_plan_eau <lgl>,
#> #   uri_entite_hydro_plan_eau <lgl>, code_mer_ocean <lgl>,
#> #   ressource_cont_non_referencee <lgl>,
#> #   ressource_cont_non_referencee_info <lgl>, code_point_referent <chr>, …

# Retrieve the withdrawal time series of the devices located in Romilly-sur-Seine
get_prelevements_chroniques(code_commune_insee = "10323")
#> # A tibble: 45 × 28
#>    code_o…¹ annee volume code_…² libel…³ code_…⁴ libel…⁵ code_…⁶ libel…⁷ code_…⁸
#>    <chr>    <int>  <dbl> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 OPR0000…  2012 617624 AEP     EAU PO… 1       Contrô… 1       Correc… REA    
#>  2 OPR0000…  2013 463545 AEP     EAU PO… 1       Contrô… 1       Correc… REA    
#>  3 OPR0000…  2014 535557 AEP     EAU PO… 1       Contrô… 1       Correc… REA    
#>  4 OPR0000…  2015 541785 AEP     EAU PO… 1       Contrô… 1       Correc… REA    
#>  5 OPR0000…  2016 478938 AEP     EAU PO… 1       Contrô… 1       Correc… REA    
#>  6 OPR0000…  2017 402363 AEP     EAU PO… 1       Contrô… 1       Correc… REA    
#>  7 OPR0000…  2018 402383 AEP     EAU PO… 1       Contrô… 1       Correc… REA    
#>  8 OPR0000…  2012 617624 AEP     EAU PO… 1       Contrô… 1       Correc… REA    
#>  9 OPR0000…  2013 463545 AEP     EAU PO… 1       Contrô… 1       Correc… REA    
#> 10 OPR0000…  2014 535557 AEP     EAU PO… 1       Contrô… 1       Correc… REA    
#> # … with 35 more rows, 18 more variables: libelle_statut_instruction <chr>,
#> #   code_mode_obtention_volume <chr>, libelle_mode_obtention_volume <chr>,
#> #   prelevement_ecrasant <lgl>, producteur_donnee <chr>, longitude <dbl>,
#> #   latitude <dbl>, code_commune_insee <chr>, nom_commune <chr>,
#> #   code_departement <chr>, libelle_departement <chr>, nom_ouvrage <chr>,
#> #   geometry.type <chr>, geometry.crs.type <chr>,
#> #   geometry.crs.properties.name <chr>, geometry.coordinates1 <dbl>, …
# }