Retrieve withdrawal devices from API "Prélèvements en eau"

See the API documentation for available filter parameters: https://hubeau.eaufrance.fr/page/api-prelevements-eau

get_prelevements_ouvrages(
  params,
  cfg = config::get(file = system.file("config.yml", package = "hubeau"))
)

Arguments

params
a list the list of parameters of the queries and their values in the format list(ParamName = "Param value", ...), use the function get_available_params for a list of the available parameters for a given operation in an API and see the API documentation for the complete list of available filter parameters

cfg
a config object Configuration of the communication. Use by default the internal package configuration

Value

a tibble::tibble with all available parameters in columns and one row by device.

Examples

# Retrieve the withdrawal devices located in Romilly-sur-Seine
get_prelevements_ouvrages(list(code_commune_insee = "10323"))

#> Error: `map_df()` requires dplyr