Retrieve data from API "Qualité des nappes d'eau souterraines"

The data originate from the "ADES" database (https://ades.eaufrance.fr/). Available endpoints are:

get_qualite_nappes_stations retrieves measuring stations for groundwater quality

get_qualite_nappes_analyses retrieves the result analysis of quality measurement

See the API documentation for available filter parameters: https://hubeau.eaufrance.fr/page/api-qualite-nappes

get_qualite_nappes_analyses(...)

get_qualite_nappes_stations(...)

Arguments

...

parameters of the queries and their values in the format Param1_Name = "Param1 value", Param2_Name = "Param2 value", use the function list_params for a list of the available filter parameters for a given API endpoint and see the API documentation for their description

Value

A tibble::tibble with one row by record and one column by field.

Examples

# \dontrun{
# List of available filter parameters on 'get_qualite_nappes_stations'
list_params("qualite_nappes", "stations")
#>  [1] "bassin_dce"                           
#>  [2] "bbox"                                 
#>  [3] "bss_id"                               
#>  [4] "circonscription_administrative_bassin"
#>  [5] "code_commune"                         
#>  [6] "code_entite_hg_bdlisa"                
#>  [7] "code_masse_eau_edl"                   
#>  [8] "code_masse_eau_rap"                   
#>  [9] "date_max_maj"                         
#> [10] "date_min_maj"                         
#> [11] "fields"                               
#> [12] "format"                               
#> [13] "nom_entite_hg_bdlisa"                 
#> [14] "nom_masse_eau_edl"                    
#> [15] "nom_masse_eau_rap"                    
#> [16] "nom_region"                           
#> [17] "nom_reseau"                           
#> [18] "num_departement"                      
#> [19] "page"                                 
#> [20] "prof_invest_max"                      
#> [21] "prof_invest_min"                      
#> [22] "size"                                 
#> [23] "srid"                                 

# List of stations available in Hérault department
get_qualite_nappes_stations(code_commune = 34116)
#> # A tibble: 6 × 59
#>   bss_id     code_bss          urn_bss         date_debut_mesure date_fin_mesure
#>   <chr>      <chr>             <chr>           <chr>             <chr>          
#> 1 BSS002GNSA 09903X0125/SEC    http://service… 2000-03-14        2023-10-16     
#> 2 BSS002GNNH 09903X0029/S      http://service… NA                NA             
#> 3 BSS002GNQP 09903X0084/SOURCE http://service… 1996-12-17        2023-05-26     
#> 4 BSS002GPFT 09906X0162/SEC    http://service… NA                NA             
#> 5 BSS002GNRW 09903X0121/MNTLET http://service… NA                NA             
#> 6 BSS002GNQZ 09903X0094/F2GRAB http://service… 1996-12-17        2023-08-31     
#> # ℹ 54 more variables: precision_coordonnees <lgl>, longitude <dbl>,
#> #   latitude <dbl>, altitude <chr>, code_insee <chr>, nom_commune <chr>,
#> #   num_departement <chr>, nom_departement <chr>, nom_region <chr>,
#> #   circonscriptions_administrative_bassin <chr>, bassin_dce <chr>,
#> #   urn_bassin_dce <chr>, code_nature_pe <int>, nom_nature_pe <chr>,
#> #   uri_nature_pe <chr>, libelle_pe <chr>, code_caracteristique_aquifere <chr>,
#> #   nom_caracteristique_aquifere <chr>, uri_caracteristique_aquifere <chr>, …

# List of available filter parameters on 'get_qualite_nappes_analyses'
list_params("qualite_nappes", "analyses")
#>  [1] "bbox"                                      
#>  [2] "bss_id"                                    
#>  [3] "code_bassin_dce"                           
#>  [4] "code_circonscription_administrative_bassin"
#>  [5] "code_entite_hg_bdlisa"                     
#>  [6] "code_fraction"                             
#>  [7] "code_groupe_parametre"                     
#>  [8] "code_insee_actuel"                         
#>  [9] "code_lieu_analyse"                         
#> [10] "code_masse_eau_edl"                        
#> [11] "code_masse_eau_rap"                        
#> [12] "code_methode"                              
#> [13] "code_param"                                
#> [14] "code_producteur"                           
#> [15] "code_qualification"                        
#> [16] "code_region"                               
#> [17] "code_remarque_analyse"                     
#> [18] "code_reseau"                               
#> [19] "code_statut_analyse"                       
#> [20] "code_type_point_eau"                       
#> [21] "code_type_qualito"                         
#> [22] "code_unite"                                
#> [23] "date_debut_prelevement"                    
#> [24] "date_fin_prelevement"                      
#> [25] "date_max_maj"                              
#> [26] "date_min_maj"                              
#> [27] "fields"                                    
#> [28] "nom_bassin_dce"                            
#> [29] "nom_circonscription_administrative_bassin" 
#> [30] "nom_commune_actuel"                        
#> [31] "nom_departement"                           
#> [32] "nom_entite_hg_bdlisa"                      
#> [33] "nom_fraction"                              
#> [34] "nom_groupe_parametre"                      
#> [35] "nom_lieu_analyse"                          
#> [36] "nom_masse_eau_edl"                         
#> [37] "nom_masse_eau_rap"                         
#> [38] "nom_methode"                               
#> [39] "nom_qualification"                         
#> [40] "nom_region"                                
#> [41] "nom_remarque_analyse"                      
#> [42] "nom_reseau"                                
#> [43] "nom_statut_analyse"                        
#> [44] "nom_type_point_eau"                        
#> [45] "nom_type_qualito"                          
#> [46] "nom_unite"                                 
#> [47] "num_departement"                           
#> [48] "page"                                      
#> [49] "resultat_max"                              
#> [50] "resultat_min"                              
#> [51] "size"                                      
#> [52] "sort"                                      

# Get results of analysis realised at Grabels in 2019
get_qualite_nappes_analyses(bss_id = "BSS002GNSA",
                            date_debut_prelevement = "2019-11-12")
#> # A tibble: 862 × 180
#>    bss_id     code_bss urn_bss precision_coordonnees longitude latitude altitude
#>    <chr>      <chr>    <chr>   <lgl>                     <dbl>    <dbl> <chr>   
#>  1 BSS002GNSA 09903X0… http:/… NA                         3.80     43.7 104.0   
#>  2 BSS002GNSA 09903X0… http:/… NA                         3.80     43.7 104.0   
#>  3 BSS002GNSA 09903X0… http:/… NA                         3.80     43.7 104.0   
#>  4 BSS002GNSA 09903X0… http:/… NA                         3.80     43.7 104.0   
#>  5 BSS002GNSA 09903X0… http:/… NA                         3.80     43.7 104.0   
#>  6 BSS002GNSA 09903X0… http:/… NA                         3.80     43.7 104.0   
#>  7 BSS002GNSA 09903X0… http:/… NA                         3.80     43.7 104.0   
#>  8 BSS002GNSA 09903X0… http:/… NA                         3.80     43.7 104.0   
#>  9 BSS002GNSA 09903X0… http:/… NA                         3.80     43.7 104.0   
#> 10 BSS002GNSA 09903X0… http:/… NA                         3.80     43.7 104.0   
#> # ℹ 852 more rows
#> # ℹ 173 more variables: code_insee_actuel <chr>, nom_commune_actuel <chr>,
#> #   num_departement <chr>, nom_departement <chr>, code_region <chr>,
#> #   nom_region <chr>, code_circonscription_administrative_bassin <chr>,
#> #   nom_circonscription_administrative_bassin <chr>, code_bassin_dce <chr>,
#> #   nom_bassin_dce <chr>, urn_bassin_dce <chr>, code_type_point_eau <int>,
#> #   nom_type_point_eau <chr>, codes_entite_hg_bdlisa <chr>, …
# }