Retrieve data from API "Température des cours d'eau"

Available endpoints are:

get_temperature_stations retrieves temperature stations

get_temperature_chronique retrieves temperature time series

See the API documentation of each endpoint for available filter parameters: https://hubeau.eaufrance.fr/page/api-temperature-continu

get_temperature_stations(params)

get_temperature_chronique(params)

Arguments

params

a list the list of parameters of the queries and their values in the format list(ParamName = "Param value", ...), use the function list_params for a list of the available filter parameters for a given API endpoint and see the API documentation for their description

Value

A tibble::tibble with one row by record and one column by field.

Examples

# \dontrun{
# Retrieve the temperature stations in the department of Loiret
get_temperature_stations(list(code_departement = "45"))
#> # A tibble: 25 × 43
#>    code_station libell…¹ uri_s…² local…³ coord…⁴ coord…⁵ code_…⁶ longi…⁷ latit…⁸
#>    <chr>        <chr>    <chr>   <chr>     <dbl>   <dbl>   <int>   <dbl>   <dbl>
#>  1 04051650     ARDOUX … https:… LIEU-D… 608980. 6746966      26    1.78    47.8
#>  2 04051650     ARDOUX … https:… LIEU-D… 608980. 6746966      26    1.78    47.8
#>  3 04051650     ARDOUX … https:… LIEU-D… 608980. 6746966      26    1.78    47.8
#>  4 04051650     ARDOUX … https:… LIEU-D… 608980. 6746966      26    1.78    47.8
#>  5 04051650     ARDOUX … https:… LIEU-D… 608980. 6746966      26    1.78    47.8
#>  6 04051125     DHUY à … https:… PONT D… 628798. 6746262      26    2.05    47.8
#>  7 04051125     DHUY à … https:… PONT D… 628798. 6746262      26    2.05    47.8
#>  8 04051125     DHUY à … https:… PONT D… 628798. 6746262      26    2.05    47.8
#>  9 04051125     DHUY à … https:… PONT D… 628798. 6746262      26    2.05    47.8
#> 10 04051125     DHUY à … https:… PONT D… 628798. 6746262      26    2.05    47.8
#> # … with 15 more rows, 34 more variables: code_commune <chr>,
#> #   libelle_commune <chr>, code_departement <chr>, libelle_departement <chr>,
#> #   code_region <chr>, libelle_region <chr>, code_troncon_hydro <chr>,
#> #   code_cours_eau <chr>, libelle_cours_eau <lgl>, uri_cours_eau <chr>,
#> #   code_masse_eau <chr>, libelle_masse_eau <lgl>, uri_masse_eau <chr>,
#> #   code_sous_bassin <lgl>, libelle_sous_bassin <lgl>, code_bassin <lgl>,
#> #   libelle_bassin <lgl>, uri_bassin <lgl>, pk <dbl>, altitude <dbl>, …

# Retrieve the temperature from 2012-01-01 to 2012-01-05 at site 04051125

get_temperature_chronique(list(
  code_station = "04051125",
  date_debut_mesure = "2012-01-01",
  date_fin_mesure="2012-01-05",
  fields = paste("code_station,date_mesure_temp,heure_mesure_temp,resultat,symbole_unite")
))
#> # A tibble: 120 × 5
#>    code_station date_mesure_temp heure_mesure_temp resultat symbole_unite
#>    <chr>        <chr>            <chr>                <dbl> <chr>        
#>  1 04051125     2012-01-01       00:00:00              8.67 °C           
#>  2 04051125     2012-01-01       01:00:00              8.72 °C           
#>  3 04051125     2012-01-01       02:00:00              8.77 °C           
#>  4 04051125     2012-01-01       03:00:00              8.84 °C           
#>  5 04051125     2012-01-01       04:00:00              8.84 °C           
#>  6 04051125     2012-01-01       05:00:00              8.79 °C           
#>  7 04051125     2012-01-01       06:00:00              8.77 °C           
#>  8 04051125     2012-01-01       07:00:00              8.77 °C           
#>  9 04051125     2012-01-01       08:00:00              8.74 °C           
#> 10 04051125     2012-01-01       09:00:00              8.79 °C           
#> # … with 110 more rows

# }