List available Hub'Eau APIs, endpoints and filter parameters

list_apis() returns the list of available APIs in the package.

list_endpoints() returns the list of available endpoints for an API.

list_params() returns the list of available parameters for an API endpoint.

list_apis()

list_endpoints(api)

list_params(api, endpoint)

Arguments

api

a character name of the API (e.g.: "indicateurs_services", "prelevements"...), see example for available APIs

endpoint

a character name of the endpoint, see example for available endpoints in an API

Value

character vector of APIs, endpoints or filter parameters

Details

The listed APIs correspond to the term [API] used in the name of the functions get_[API]_[endpoint] used for querying the APIs.

Examples

# To get the available APIs in the package
list_apis()
#>  [1] "prelevements"         "indicateurs_services" "hydrometrie"         
#>  [4] "niveaux_nappes"       "poisson"              "ecoulement"          
#>  [7] "hydrobio"             "temperature"          "qualite_eau_potable" 
#> [10] "qualite_nappes"       "qualite_rivieres"    

# To get the available endpoints in an API
list_endpoints("prelevements")
#> [1] "chroniques"         "ouvrages"           "points_prelevement"

# To get available parameters in endpoint "chroniques" of the API "prelevements"
list_params(api = "prelevements", endpoint = "chroniques")
#>  [1] "annee"                      "bbox"                      
#>  [3] "code_commune_insee"         "code_departement"          
#>  [5] "code_mode_obtention_volume" "code_ouvrage"              
#>  [7] "code_qualification_volume"  "code_statut_instruction"   
#>  [9] "code_statut_volume"         "code_usage"                
#> [11] "distance"                   "fields"                    
#> [13] "format"                     "latitude"                  
#> [15] "libelle_departement"        "longitude"                 
#> [17] "nom_commune"                "page"                      
#> [19] "prelevement_ecrasant"       "producteur_donnee"         
#> [21] "size"                       "sort"                      
#> [23] "volume_max"                 "volume_min"