Retrieve data from API "Ecoulement des cours d'eau"

The data originate from the "ONDE" river low waters monitoring network. Available endpoints are:

get_ecoulement_stations retrieves site data and locations

get_ecoulement_observations retrieves flow information

get_ecoulement_campagnes retrieves annual surveys

See the API documentation for available filter parameters: https://hubeau.eaufrance.fr/page/api-ecoulement

get_ecoulement_stations(params)

get_ecoulement_observations(params)

get_ecoulement_campagnes(params)

Arguments

params

a list the list of parameters of the queries and their values in the format list(ParamName = "Param value", ...), use the function list_params for a list of the available filter parameters for a given API endpoint and see the API documentation for their description

Value

A tibble::tibble with one row by record and one column by field.

Examples

# \dontrun{
# Retrieve 2022 observation campaigns in the Jura French department
get_ecoulement_campagnes(
  list(code_departement = "39",
       date_campagne_min = "2022-01-01",
       date_campagne_max = "2022-12-31")
)
#> # A tibble: 8 × 10
#>   code_campagne date_c…¹ nombr…² code_…³ libel…⁴ code_…⁵ libel…⁶ uri_r…⁷ code_…⁸
#>           <int> <chr>      <int>   <int> <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1         96881 2022-10…       5       2 complé… 000000… Onde O… http:/… 39     
#> 2         96802 2022-09…       5       1 usuelle 000000… Onde O… http:/… 39     
#> 3         96730 2022-09…       5       2 complé… 000000… Onde O… http:/… 39     
#> 4         96655 2022-08…       5       1 usuelle 000000… Onde O… http:/… 39     
#> 5         96594 2022-08…       5       2 complé… 000000… Onde O… http:/… 39     
#> 6         96509 2022-07…       5       1 usuelle 000000… Onde O… http:/… 39     
#> 7         96335 2022-06…       5       1 usuelle 000000… Onde O… http:/… 39     
#> 8         96137 2022-05…       5       1 usuelle 000000… Onde O… http:/… 39     
#> # … with 1 more variable: libelle_departement <chr>, and abbreviated variable
#> #   names ¹​date_campagne, ²​nombre_modalite_ecoulement, ³​code_type_campagne,
#> #   ⁴​libelle_type_campagne, ⁵​code_reseau, ⁶​libelle_reseau, ⁷​uri_reseau,
#> #   ⁸​code_departement

# Retrieve river stations
stations_39 <- get_ecoulement_stations(
  list(code_departement = "39",
       fields = "code_station,libelle_cours_eau,libelle_commune")
)
stations_39
#> # A tibble: 43 × 3
#>    code_station libelle_commune libelle_cours_eau    
#>    <chr>        <chr>           <chr>                
#>  1 U1084011     OFFLANGES       Ruisseau de la Vèze  
#>  2 U1084012     SALIGNEY        NA                   
#>  3 U1120011     MENOTEY         Ruisseau de Borne    
#>  4 U2530001     EVANS           La Grabusse          
#>  5 U2530002     FRAISANS        La Morte             
#>  6 U2542011     AMANGE          Ruisseau Vèze        
#>  7 U2542012     CHATENOIS       Ruisseau des Vaux    
#>  8 U2542013     LOUVATANGE      L'Arne               
#>  9 U2639011     CHAMBLAY        La Larine            
#> 10 U2640002     CHAMBLAY        Ruisseau de Clairvent
#> # … with 33 more rows

# Get the query parameters for the requested API/endpoint
list_params(api = "ecoulement",
                     endpoint = "observations")
#>  [1] "format"               "code_station"         "libelle_station"     
#>  [4] "code_departement"     "libelle_departement"  "code_commune"        
#>  [7] "libelle_commune"      "code_region"          "libelle_region"      
#> [10] "code_bassin"          "libelle_bassin"       "code_cours_eau"      
#> [13] "libelle_cours_eau"    "code_campagne"        "code_reseau"         
#> [16] "libelle_reseau"       "date_observation_min" "date_observation_max"
#> [19] "code_ecoulement"      "libelle_ecoulement"   "longitude"           
#> [22] "latitude"             "distance"             "bbox"                
#> [25] "Numéro de la page"    "size"                 "fields"              
#> [28] "Accept"              

# Retrieve the river flow data in the Jura departement in 2022 with
# a selection of the fields
onde_39 <- get_ecoulement_observations(
  list(code_departement = "39",
       date_observation_min = "2022-01-01",
       date_observation_max = "2022-12-31",
       fields = "code_station,libelle_station,date_observation,libelle_ecoulement")
)
onde_39
#> # A tibble: 336 × 4
#>    code_station libelle_station                   date_observation libelle_eco…¹
#>    <chr>        <chr>                             <chr>            <chr>        
#>  1 U2640003     Pont D7 à GERMIGNEY               2022-10-24       Ecoulement n…
#>  2 U2655034     ruisseau de la verine à Vaudrey   2022-10-24       Ecoulement v…
#>  3 U2640002     ruisseau de clairvent à Chamblay  2022-10-24       Ecoulement v…
#>  4 U1084012     Ruisseau de la Serre à Saligney   2022-10-24       Ecoulement v…
#>  5 U1120011     ruisseau de la foret à Menotey    2022-10-24       Ecoulement v…
#>  6 U2722011     Rivière la Sablonne à Bousselange 2022-10-24       Ecoulement v…
#>  7 U2530001     La grabusse à Dampierre           2022-10-24       Ecoulement v…
#>  8 U2542013     Ruisseau l'Arne à Louvatange      2022-10-24       Ecoulement v…
#>  9 U2542012     Ruisseau des Vaux à Chatenois     2022-10-24       Ecoulement v…
#> 10 U2542011     Ruisseau Du Moulinot à Amange     2022-10-24       Ecoulement v…
#> # … with 326 more rows, and abbreviated variable name ¹​libelle_ecoulement
# }