Retrieve data from API "Hydrobiologie"

The data originate from the "NAIADES" database. Available endpoints are:

get_hydrobio_stations retrieves site data and locations

get_hydrobio_indices retrieves bioassessment indices values

get_hydrobio_taxons retrieves taxa data

See the API documentation for available filter parameters: https://hubeau.eaufrance.fr/page/api-hydrobiologie

get_hydrobio_stations(params)

get_hydrobio_indices(params)

get_hydrobio_taxons(params)

Arguments

params

a list the list of parameters of the queries and their values in the format list(ParamName = "Param value", ...), use the function list_params for a list of the available filter parameters for a given API endpoint and see the API documentation for their description

Value

A tibble::tibble with one row by record and one column by field.

Examples

if(interactive()) {
# Retrieve the hydrobiology monitoring sites in the Pays-de-Loire region
list_params(api = "hydrobio",
            endpoint = "stations")

stations_pdl <- get_hydrobio_stations(
 list(code_region = 52))

# Retrieve the hydrobiological bioassessment indices in the city of Rennes
list_params(api = "hydrobio",
endpoint = "indices")

indices_rennes <- get_hydrobio_indices(
 list(code_commune = 35051))

# species records in the city of Rennes
list_params(api = "hydrobio",
            endpoint = "taxons")

taxons_rennes <- hubeau::get_hydrobio_taxons(
 list(code_commune = 35051))
}