Retrieve data from API "Indicateurs des services"

Retrieve performance indicators collected in drinking water and sanitation services in France.

See list_params and the API documentation for available filter parameters: https://hubeau.eaufrance.fr/page/api-indicateurs-services

get_indicateurs_services_communes(...)

get_indicateurs_services_indicateurs(...)

get_indicateurs_services_services(...)

Arguments

...

parameters of the queries and their values in the format Param1_Name = "Param1 value", Param2_Name = "Param2 value", use the function list_params for a list of the available filter parameters for a given API endpoint and see the API documentation for their description

Value

get_indicateurs_services_communes returns a tibble::tibble with one row by commune, by service and by year and the following columns:

"code_commune_insee": character identifier of the commune

"nom_commune": character name of the commune

"codes_service": integer identifier of the drinking water supply and/or sanitation service

"annee": integer year of the data

The following columns are the performance indicators flagged by their respective codes. The documentation of these codes can be found at this URL: https://www.services.eaufrance.fr/indicateurs/indicateurs.

get_indicateurs_services_indicateurs returns a tibble::tibble with one row by service and by year and the following columns:

"code_service": character identifier of the service

"nom_service": character name of the service

"numero_siren ": character SIREN identifier of the service

"type_collectivite": character kind of community

"mode_gestion": character management mechanism of the service

"annee": integer year of the data

"indicateur": value of the indicator

"uri_indicateur": the link to the indicator documentation

get_indicateurs_services_services returns a tibble::tibble with one row by commune, by service and by year and the following columns:

"code_service": character identifier of the service

"nom_service": character name of the service

"code_commune_insee": character identifier of the commune

"nom_commune": character name of the commune

"numero_siren ": character SIREN identifier of the service

"type_collectivite": character kind of community

"mode_gestion": character management mechanism of the service

"annee": integer year of the data

The following columns are the performance indicators flagged by their respective codes. The documentation of these codes can be found at this URL: https://www.services.eaufrance.fr/indicateurs/indicateurs.

Examples

# \dontrun{
# Retrieve performance indicator time series in the commune of Romilly-sur-Seine
get_indicateurs_services_communes(code_commune = "10323")
#> # A tibble: 28 × 39
#>    code_commune_insee nom_commune codes_service noms_service annee D102.0 D151.0
#>    <chr>              <chr>               <int> <chr>        <int>  <dbl>  <dbl>
#>  1 10323              Romilly-su…        148782 eau potable   2018   1.91      1
#>  2 10323              Romilly-su…        148782 eau potable   2017   2.07      1
#>  3 10323              NA                 148783 assainissem…  2017  NA        NA
#>  4 10323              Romilly-su…        148783 assainissem…  2017  NA        NA
#>  5 10323              Romilly-su…        148782 eau potable   2016   2.18      1
#>  6 10323              Romilly-su…        148783 assainissem…  2016  NA        NA
#>  7 10323              NA                 148783 assainissem…  2016  NA        NA
#>  8 10323              Romilly-su…        148782 eau potable   2015   2.2       1
#>  9 10323              NA                 148783 assainissem…  2015  NA        NA
#> 10 10323              Romilly-su…        148783 assainissem…  2015  NA        NA
#> # ℹ 18 more rows
#> # ℹ 32 more variables: P101.1 <dbl>, P101.1a <lgl>, P101.1b <lgl>,
#> #   P102.1 <dbl>, P102.1a <lgl>, P102.1b <lgl>, P104.3 <dbl>, P106.3 <dbl>,
#> #   P107.2 <dbl>, P108.3 <dbl>, P151.1 <dbl>, P152.1 <dbl>, P155.1 <dbl>,
#> #   VP.177 <dbl>, VP.190 <dbl>, VP.178 <dbl>, VP.191 <dbl>, VP.179 <dbl>,
#> #   VP.061 <dbl>, DC.192 <dbl>, VP.228 <dbl>, VP.231 <dbl>, D204.0 <dbl>,
#> #   P201.1 <dbl>, P203.3 <dbl>, P204.3 <dbl>, P205.3 <dbl>, P258.1 <dbl>, …

# Retrieve the drinking water withdrawal indicators of the year 2012 for all services
get_indicateurs_services_indicateurs(code_indicateur = "D102.0", annee = "2012")
#> # A tibble: 5,471 × 8
#>    code_service nom_service numero_siren type_collectivite    mode_gestion annee
#>           <int> <chr>       <chr>        <chr>                <chr>        <int>
#>  1        99941 eau potable 258803303    Syndicat Intercommu… Régie         2012
#>  2        99919 eau potable NA           Syndicat Intercommu… Régie         2012
#>  3        99917 eau potable 258800838    Syndicat Intercommu… Délégation    2012
#>  4        99916 eau potable 258800028    Syndicat Intercommu… Régie         2012
#>  5        99915 eau potable 258800275    Syndicat Intercommu… Régie         2012
#>  6        99914 eau potable 248800237    Syndicat Intercommu… Régie         2012
#>  7        99913 eau potable 258800911    Syndicat Intercommu… Régie         2012
#>  8        99911 eau potable 258803048    Syndicat Intercommu… Régie         2012
#>  9        99910 eau potable 258800044    Syndicat Intercommu… Régie         2012
#> 10        99909 eau potable 258800713    Syndicat Intercommu… Régie         2012
#> # ℹ 5,461 more rows
#> # ℹ 2 more variables: indicateur <dbl>, uri_indicateur <chr>

# Retrieve performance indicator time series of Romilly-sur-Seine with service details
get_indicateurs_services_services(code_commune = "10323")
#> # A tibble: 9 × 38
#>   code_service nom_service codes_commune noms_commune      numero_siren
#>          <int> <chr>       <chr>         <chr>             <chr>       
#> 1       148782 eau potable 10323         Romilly-sur-Seine 211003124   
#> 2       148782 eau potable 10323         Romilly-sur-Seine 211003124   
#> 3       148782 eau potable 10323         Romilly-sur-Seine 211003124   
#> 4       148782 eau potable 10323         Romilly-sur-Seine 211003124   
#> 5       148782 eau potable 10323         Romilly-sur-Seine 211003124   
#> 6       148782 eau potable 10323         Romilly-sur-Seine 211003124   
#> 7       148782 eau potable 10323         Romilly-sur-Seine 211003124   
#> 8       148782 eau potable 10323         Romilly-sur-Seine 211003124   
#> 9       148782 eau potable 10323         Romilly-sur-Seine 211003124   
#> # ℹ 33 more variables: type_collectivite <chr>, mode_gestion <chr>,
#> #   annee <int>, D101.0 <lgl>, D102.0 <dbl>, D151.0 <dbl>, P101.1 <dbl>,
#> #   P101.1a <lgl>, P101.1b <lgl>, P102.1 <dbl>, P102.1a <lgl>, P102.1b <lgl>,
#> #   P104.3 <dbl>, P106.3 <dbl>, P107.2 <dbl>, P108.3 <dbl>, P151.1 <dbl>,
#> #   P152.1 <dbl>, P155.1 <dbl>, VP.056 <lgl>, VP.177 <dbl>, VP.190 <dbl>,
#> #   VP.178 <dbl>, VP.191 <dbl>, VP.179 <dbl>, VP.059 <lgl>, VP.060 <lgl>,
#> #   VP.061 <dbl>, VP.063 <lgl>, VP.201 <lgl>, DC.192 <dbl>, VP.228 <dbl>, …
# }