Retrieve data from API "Prélèvements en eau"

Available endpoints are:

get_prelevements_points_prelevement retrieves withdrawal points

get_prelevements_ouvrages retrieves withdrawal devices

get_prelevements_chroniques retrieves time series of withdrawals

See the API documentation for available filter parameters: https://hubeau.eaufrance.fr/page/api-prelevements-eau

get_prelevements_points_prelevement(...)

get_prelevements_ouvrages(...)

get_prelevements_chroniques(...)

Arguments

...

parameters of the queries and their values in the format Param1_Name = "Param1 value", Param2_Name = "Param2 value", use the function list_params for a list of the available filter parameters for a given API endpoint and see the API documentation for their description

Value

A tibble::tibble with one row by record and one column by field.

Examples

# \dontrun{
# Retrieve the withdrawal points located in Romilly-sur-Seine
get_prelevements_points_prelevement(code_commune_insee = "10323")
#> # A tibble: 9 × 28
#>   code_point_prelevement nom_point_prelevement            date_exploitation_de…¹
#>   <chr>                  <chr>                            <chr>                 
#> 1 PTP000000000047369     NA                               1987-01-01            
#> 2 PTP000000000047370     ROMILLY SUR SEINE                2008-01-18            
#> 3 PTP000000000053334     NA                               1900-01-01            
#> 4 PTP000000000054886     S.N.C.F                          2006-09-28            
#> 5 PTP000000000054887     S.N.C.F                          2006-09-28            
#> 6 PTP000000000233572     SCEA FERME DES HAUTS BUISSONS- … 1992-01-01            
#> 7 PTP000000000434018     Soc des Carrières de l'Est - R … 2014-01-01            
#> 8 PTP000000000762434     NA                               1900-01-01            
#> 9 PTP000000000788799     CRBPE ROMILLY SUR SEINE          2014-01-01            
#> # ℹ abbreviated name: ¹​date_exploitation_debut
#> # ℹ 25 more variables: date_exploitation_fin <lgl>, code_type_milieu <chr>,
#> #   libelle_type_milieu <chr>, code_nature <chr>, libelle_nature <chr>,
#> #   lieu_dit <lgl>, commentaire <lgl>, code_commune_insee <chr>,
#> #   nom_commune <chr>, code_departement <chr>, libelle_departement <chr>,
#> #   code_entite_hydro_cours_eau <lgl>, uri_entite_hydro_cours_eau <lgl>,
#> #   code_entite_hydro_plan_eau <lgl>, uri_entite_hydro_plan_eau <lgl>, …

# Retrieve the withdrawal devices located in Romilly-sur-Seine
get_prelevements_ouvrages(code_commune_insee = "10323")
#> # A tibble: 14 × 33
#>    code_ouvrage  nom_ouvrage             id_local_ouvrage date_exploitation_de…¹
#>    <chr>         <chr>                   <chr>            <chr>                 
#>  1 OPR0000032603 ROMILLY SUR SEINE       02612X1004/PAEP… 1987-01-01            
#>  2 OPR0000032604 ROMILLY SUR SEINE       02612X1011/FAEP… 2008-01-18            
#>  3 OPR0000038567 SCEA FERME DES HAUTS B… BSS95456G_018    1900-01-01            
#>  4 OPR0000040119 S.N.C.F                 BSS00593V_018    2006-09-28            
#>  5 OPR0000040120 S.N.C.F                 BSS00594W_018    2006-09-28            
#>  6 OPR0000199638 SCEA FERME DES HAUTS B… BSS95455F_018    1992-01-01            
#>  7 OPR0000333915 Soc des Carrières de l… 32055_017        2014-01-01            
#>  8 OPR0000586224 SCEA DU MERISIER-10000… BSS95299L_018    1900-01-01            
#>  9 OPR0000608675 CRBPE ROMILLY SUR SEINE 31132_018        2014-01-01            
#> 10 OPR0000618312 ROMILLY SUR SEINE       02612X1011/FAEP… 2008-01-18            
#> 11 OPR0000620041 CRBPE ROMILLY SUR SEINE 31132_018        2014-01-01            
#> 12 OPR0000621348 S.N.C.F                 BSS00594W_018    2006-09-28            
#> 13 OPR0000627660 S.N.C.F                 BSS00593V_018    2006-09-28            
#> 14 OPR0000627936 SCEA FERME DES HAUTS B… BSS95455F_018    1992-01-01            
#> # ℹ abbreviated name: ¹​date_exploitation_debut
#> # ℹ 29 more variables: date_exploitation_fin <lgl>, code_precision_coord <chr>,
#> #   libelle_precision_coord <chr>, commentaire <lgl>, code_commune_insee <chr>,
#> #   nom_commune <chr>, code_departement <chr>, libelle_departement <chr>,
#> #   code_type_milieu <chr>, libelle_type_milieu <chr>,
#> #   code_entite_hydro_cours_eau <lgl>, uri_entite_hydro_cours_eau <lgl>,
#> #   code_entite_hydro_plan_eau <lgl>, uri_entite_hydro_plan_eau <lgl>, …

# Retrieve the withdrawal time series of the devices located in Romilly-sur-Seine
get_prelevements_chroniques(code_commune_insee = "10323")
#> # A tibble: 52 × 28
#>    code_ouvrage  annee volume code_usage libelle_usage code_statut_volume
#>    <chr>         <int>  <dbl> <chr>      <chr>         <chr>             
#>  1 OPR0000032603  2012 617624 AEP        EAU POTABLE   1                 
#>  2 OPR0000032603  2013 463545 AEP        EAU POTABLE   1                 
#>  3 OPR0000032603  2014 535557 AEP        EAU POTABLE   1                 
#>  4 OPR0000032603  2015 541785 AEP        EAU POTABLE   1                 
#>  5 OPR0000032603  2016 478938 AEP        EAU POTABLE   1                 
#>  6 OPR0000032603  2017 402363 AEP        EAU POTABLE   1                 
#>  7 OPR0000032603  2018 402383 AEP        EAU POTABLE   1                 
#>  8 OPR0000032604  2012 617624 AEP        EAU POTABLE   1                 
#>  9 OPR0000032604  2013 463545 AEP        EAU POTABLE   1                 
#> 10 OPR0000032604  2014 535557 AEP        EAU POTABLE   1                 
#> # ℹ 42 more rows
#> # ℹ 22 more variables: libelle_statut_volume <chr>,
#> #   code_qualification_volume <chr>, libelle_qualification_volume <chr>,
#> #   code_statut_instruction <chr>, libelle_statut_instruction <chr>,
#> #   code_mode_obtention_volume <chr>, libelle_mode_obtention_volume <chr>,
#> #   prelevement_ecrasant <lgl>, producteur_donnee <chr>, longitude <dbl>,
#> #   latitude <dbl>, code_commune_insee <chr>, nom_commune <chr>, …
# }