Retrieve data from API "Qualité de l'eau potable"

Results of the sanitary control of the distributed water commune by commune: samples and results of the analyses carried out within the framework of the regulatory sanitary control on the distribution units or the installations directly upstream, and links between communes and distribution units. The elements made available in this dataset correspond to a compilation of analysis bulletins published online, commune by commune, on the website of the Ministry of Health: https://sante.gouv.fr/sante-et-environnement/eaux/eau

Available endpoints are:

get_qualite_eau_potable_communes_udi retrieves links between "UDI" (Distribution units or networks) and communes

get_qualite_eau_potable_resultats_dis retrieves samples, analysis results and sanitary conclusions from the sanitary control of the distributed water commune by commune

See the API documentation for available filter parameters: https://hubeau.eaufrance.fr/page/api-qualite-eau-potable

get_qualite_eau_potable_communes_udi(...)

get_qualite_eau_potable_resultats_dis(...)

Arguments

...

parameters of the queries and their values in the format Param1_Name = "Param1 value", Param2_Name = "Param2 value", use the function list_params for a list of the available filter parameters for a given API endpoint and see the API documentation for their description

Value

A tibble::tibble with one row by record and one column by field.

Examples

# \dontrun{
# List of available filter parameters on 'get_qualite_eau_potable_communes_udi'
list_params("qualite_eau_potable", "communes_udi")
#> [1] "annee"        "code_commune" "code_reseau"  "fields"       "nom_commune" 
#> [6] "nom_reseau"   "page"         "size"         "sort"        

# List of UDIs available in 2022 at Grabels (INSEE code 34116)
get_qualite_eau_potable_communes_udi(annee = 2022, code_commune = 34116)
#> # A tibble: 3 × 7
#>   code_commune nom_commune nom_quartier  code_reseau nom_reseau debut_alim annee
#>   <chr>        <chr>       <chr>         <chr>       <chr>      <chr>      <chr>
#> 1 34116        GRABELS     village       034000205   MMM-GRABE… 2010-11-06 2022 
#> 2 34116        GRABELS     -             034001317   MMM-MONTP… 2016-06-15 2022 
#> 3 34116        GRABELS     Goule de Lav… 034001349   MMM-GRABE… 2010-11-06 2022 

# List of available filter parameters on 'get_qualite_eau_potable_resultats_dis'
list_params("qualite_eau_potable", "resultats_dis")
#>  [1] "borne_inf_resultat"                    
#>  [2] "borne_sup_resultat"                    
#>  [3] "code_commune"                          
#>  [4] "code_departement"                      
#>  [5] "code_lieu_analyse"                     
#>  [6] "code_parametre"                        
#>  [7] "code_parametre_cas"                    
#>  [8] "code_parametre_se"                     
#>  [9] "code_prelevement"                      
#> [10] "code_reseau"                           
#> [11] "conformite_limites_bact_prelevement"   
#> [12] "conformite_limites_pc_prelevement"     
#> [13] "conformite_references_bact_prelevement"
#> [14] "conformite_references_pc_prelevement"  
#> [15] "date_max_prelevement"                  
#> [16] "date_min_prelevement"                  
#> [17] "fields"                                
#> [18] "libelle_parametre"                     
#> [19] "libelle_parametre_maj"                 
#> [20] "nom_commune"                           
#> [21] "nom_distributeur"                      
#> [22] "nom_moa"                               
#> [23] "page"                                  
#> [24] "size"                                  
#> [25] "sort"                                  

# Get results of analysis realised at Grabels in 2022
get_qualite_eau_potable_resultats_dis(code_commune = 34116,
                                      date_min_prelevement = "2000-01-01",
                                      date_max_prelevement = "2022-12-31")
#> # A tibble: 8,246 × 34
#>    code_departement nom_departement code_prelevement code_parametre
#>    <chr>            <chr>           <chr>            <chr>         
#>  1 34               Hérault         03400287942      1303          
#>  2 34               Hérault         03400287942      1447          
#>  3 34               Hérault         03400287942      1335          
#>  4 34               Hérault         03400287942      1370          
#>  5 34               Hérault         03400287942      5901          
#>  6 34               Hérault         03400287942      1301          
#>  7 34               Hérault         03400287942      NA            
#>  8 34               Hérault         03400287942      5440          
#>  9 34               Hérault         03400287942      1302          
#> 10 34               Hérault         03400287942      1398          
#> # ℹ 8,236 more rows
#> # ℹ 30 more variables: code_parametre_se <chr>, code_parametre_cas <chr>,
#> #   libelle_parametre <chr>, libelle_parametre_maj <chr>,
#> #   libelle_parametre_web <lgl>, code_type_parametre <chr>,
#> #   code_lieu_analyse <chr>, resultat_alphanumerique <chr>,
#> #   resultat_numerique <dbl>, libelle_unite <chr>, code_unite <chr>,
#> #   limite_qualite_parametre <chr>, reference_qualite_parametre <chr>, …
# }