Retrieve data from API "Température des cours d'eau"

Available endpoints are:

get_temperature_stations retrieves temperature stations

get_temperature_chronique retrieves temperature time series

See the API documentation of each endpoint for available filter parameters: https://hubeau.eaufrance.fr/page/api-temperature-continu

get_temperature_stations(...)

get_temperature_chronique(...)

Arguments

...

parameters of the queries and their values in the format Param1_Name = "Param1 value", Param2_Name = "Param2 value", use the function list_params for a list of the available filter parameters for a given API endpoint and see the API documentation for their description

Value

A tibble::tibble with one row by record and one column by field.

Examples

# \dontrun{
# Retrieve the temperature stations in the department of Loiret
get_temperature_stations(code_departement = "45")
#> # A tibble: 5 × 47
#>   code_station libelle_station             uri_station localisation coordonnee_x
#>   <chr>        <chr>                       <chr>       <chr>               <dbl>
#> 1 04051650     ARDOUX Ã  CLERY-SAINT-ANDRE https://id… LIEU-DIT LE…      608980.
#> 2 04051125     DHUY Ã  SANDILLON           https://id… PONT D13 (R…      628798.
#> 3 04049625     BEC D'ABLE Ã  SULLY-SUR-LO… https://id… MOULIN DE R…      649238.
#> 4 04048550     NOTREURE Ã  CERNOY-EN-BERRY https://id… PONT D53 AU…      672612.
#> 5 03057000     L'OUANNE A GY-LES-NONAINS 2 https://id… Non renseig…      688537 
#> # ℹ 42 more variables: coordonnee_y <dbl>, code_type_projection <int>,
#> #   longitude <dbl>, latitude <dbl>, code_commune <chr>, libelle_commune <chr>,
#> #   code_departement <chr>, libelle_departement <chr>, code_region <chr>,
#> #   libelle_region <chr>, code_troncon_hydro <chr>, code_cours_eau <chr>,
#> #   libelle_cours_eau <chr>, uri_cours_eau <chr>, code_masse_eau <chr>,
#> #   libelle_masse_eau <chr>, uri_masse_eau <chr>, code_sous_bassin <chr>,
#> #   libelle_sous_bassin <chr>, code_bassin <chr>, libelle_bassin <chr>, …

# Retrieve the temperature from 2012-01-01 to 2012-01-05 at site 04051125

get_temperature_chronique(
  code_station = "04051125",
  date_debut_mesure = "2012-01-01",
  date_fin_mesure="2012-01-05",
  fields = paste("code_station,date_mesure_temp,heure_mesure_temp,resultat,symbole_unite")
)
#> # A tibble: 120 × 5
#>    code_station date_mesure_temp heure_mesure_temp resultat symbole_unite
#>    <chr>        <chr>            <chr>                <dbl> <chr>        
#>  1 04051125     2012-01-01       00:00:00              8.67 °C           
#>  2 04051125     2012-01-01       01:00:00              8.72 °C           
#>  3 04051125     2012-01-01       02:00:00              8.77 °C           
#>  4 04051125     2012-01-01       03:00:00              8.84 °C           
#>  5 04051125     2012-01-01       04:00:00              8.84 °C           
#>  6 04051125     2012-01-01       05:00:00              8.79 °C           
#>  7 04051125     2012-01-01       06:00:00              8.77 °C           
#>  8 04051125     2012-01-01       07:00:00              8.77 °C           
#>  9 04051125     2012-01-01       08:00:00              8.74 °C           
#> 10 04051125     2012-01-01       09:00:00              8.79 °C           
#> # ℹ 110 more rows

# }